[2180] Mechanical restructure of v2v-helper: split migrate.go and move OpenStack client
- Split migrate/migrate.go (2723 LOC) into 4 concern files within package migrate:
  - nbd_copy.go: NBD copy path + CBT sync (7 functions)
  - network.go: port reservation + OS network config (17 functions)
  - conversion.go: disk conversion pipeline + boot detection (11 functions)
  - vm_ops.go: volume lifecycle + storage setup + instance creation (22 functions)
  - migrate.go trimmed to orchestration only (6 functions, 399 LOC)
- Move pkg/utils/openstackopsutils.go → openstack/clients.go (package declaration only)
  - Update openstackops.go: remove pkg/utils import, use OpenStackClients directly
  - Update vaai_copy.go: reference openstack.CinderVolumeService
  - Move openstackopsutils_test.go → openstack/ (package openstack)
- Change Migrate.Reporter field from *reporter.Reporter to reporter.ReporterOps

Zero logic changes — all functions moved verbatim.
