+++ .specify/feature.json
@@ -1,3 +1,1 @@
-{
-  "feature_directory": "specs/2120-migration-templates"
-}
+{"feature_directory":"specs/2180-v2v-restructure"}
