+++ CLAUDE.md
@@ -163,7 +163,7 @@ kubectl -n migration-system logs <migration-name>-v2v-helper
 <!-- SPECKIT START -->
 For additional context about technologies to be used, project structure,
 shell commands, and other important information, read the current plan
-at specs/2120-migration-templates/plan.md
+at specs/2180-v2v-restructure/plan.md
 <!-- SPECKIT END -->
 
 ## graphify
