+++ v2v-helper/migrate/migrate.go
@@ -5,40 +5,22 @@ package migrate
 import (
 	"context"
 	"fmt"
-	"log"
-	"net/http"
 	"os"
 	"os/signal"
-	"regexp"
-	"strconv"
-	"strings"
 	"syscall"
 	"time"
 
-	"github.com/gophercloud/gophercloud/v2/openstack/compute/v2/flavors"
-	"github.com/gophercloud/gophercloud/v2/openstack/networking/v2/networks"
-	"github.com/gophercloud/gophercloud/v2/openstack/networking/v2/ports"
 	"github.com/pkg/errors"
-	vjailbreakv1alpha1 "github.com/platform9/vjailbreak/k8s/migration/api/v1alpha1"
 	"github.com/platform9/vjailbreak/pkg/common/constants"
-	"github.com/platform9/vjailbreak/pkg/vpwned/sdk/storage"
-	netappsdk "github.com/platform9/vjailbreak/pkg/vpwned/sdk/storage/netapp"
-	_ "github.com/platform9/vjailbreak/pkg/vpwned/sdk/storage/providers"
 	"github.com/platform9/vjailbreak/v2v-helper/nbd"
 	"github.com/platform9/vjailbreak/v2v-helper/openstack"
 	"github.com/platform9/vjailbreak/v2v-helper/pkg/k8sutils"
 	"github.com/platform9/vjailbreak/v2v-helper/pkg/utils"
-	"github.com/platform9/vjailbreak/v2v-helper/pkg/utils/vmutils"
 	"github.com/platform9/vjailbreak/v2v-helper/reporter"
 	"github.com/platform9/vjailbreak/v2v-helper/vcenter"
-	"github.com/platform9/vjailbreak/v2v-helper/virtv2v"
 	"github.com/platform9/vjailbreak/v2v-helper/vm"
-	k8stypes "k8s.io/apimachinery/pkg/types"
+	"github.com/platform9/vjailbreak/pkg/vpwned/sdk/storage"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	netutils "github.com/platform9/vjailbreak/pkg/common/utils"
-	probing "github.com/prometheus-community/pro-bing"
-	"github.com/vmware/govmomi/vim25/types"
 )
 
 type Migrate struct {
@@ -72,7 +54,7 @@ type Migrate struct {
 	ServerGroup             string
 	RDMDisks                []string
 	TenantName              string
-	Reporter                *reporter.Reporter
+	Reporter                reporter.ReporterOps
 	FallbackToDHCP          bool
 	StorageCopyMethod       string
 	// Array credentials for StorageAcceleratedCopy storage migration
@@ -134,415 +116,11 @@ const (
 	StateSyncingCBT
 )
 
-// String returns a human-readable name for the state
-func (s PeriodicSyncStates) String() string {
-	switch s {
-	case StateIdle:
-		return "Idle"
-	case StateCleaningSnapshots:
-		return "CleaningSnapshots"
-	case StateTakingSnapshot:
-		return "TakingSnapshot"
-	case StateSyncingCBT:
-		return "SyncingCBT"
-	default:
-		return "Unknown"
-	}
-}
-
 // PeriodicSyncContext holds the state machine context for periodic sync operations
 type PeriodicSyncContext struct {
 	CurrentState   PeriodicSyncStates
 	LastError      error
 	WarningMessage string // Non-empty indicates sync is in warning state (failed but will retry)
-}
-
-// disconnects the source VM's network interfaces
-func (migobj *Migrate) DisconnectSourceNetworkIfRequested() error {
-	if !migobj.DisconnectSourceNetwork {
-		return nil
-	}
-
-	migobj.logMessage(fmt.Sprintf("Disconnecting source VM network interfaces (DisconnectSourceNetwork=%v)", migobj.DisconnectSourceNetwork))
-
-	if err := migobj.VMops.DisconnectNetworkInterfaces(); err != nil {
-		errMsg := fmt.Sprintf("Failed to disconnect source VM network interfaces: %v", err)
-		migobj.logMessage("ERROR: " + errMsg)
-		return fmt.Errorf("failed to disconnect network interfaces: %w", err)
-	}
-
-	migobj.logMessage("Successfully disconnected source VM network interfaces")
-	return nil
-}
-
-func (migobj *Migrate) logMessage(message string) {
-	if migobj.InPod {
-		migobj.EventReporter <- message
-	}
-	utils.PrintLog(message)
-}
-
-// This function creates volumes in OpenStack and attaches them to the helper vm
-func (migobj *Migrate) CreateVolumes(ctx context.Context, vminfo vm.VMInfo) (vm.VMInfo, error) {
-	openstackops := migobj.Openstackclients
-	migobj.logMessage("Creating volumes in OpenStack")
-
-	for idx, vmdisk := range vminfo.VMDisks {
-		setRDMLabel := false
-		if len(vminfo.RDMDisks) > 0 {
-			setRDMLabel = true
-		}
-		volume, err := openstackops.CreateVolume(ctx, vminfo.Name+"-"+vmdisk.Name, vmdisk.Size, vminfo.OSType, vminfo.UEFI, migobj.Volumetypes[idx], setRDMLabel)
-		if err != nil {
-			return vminfo, errors.Wrap(err, "failed to create volume")
-		}
-		vminfo.VMDisks[idx].OpenstackVol = volume
-		if vminfo.VMDisks[idx].Boot {
-			err = openstackops.SetVolumeBootable(ctx, volume)
-			if err != nil {
-				return vminfo, errors.Wrap(err, "failed to set volume as bootable")
-			}
-		}
-	}
-	migobj.logMessage("Volumes created successfully")
-	return vminfo, nil
-}
-
-// applyImageMetadataForXCOPYVolumes mirrors the metadata calls that CreateVolume
-// makes on the standard (non-accelerated) path.
-func (migobj *Migrate) applyImageMetadataForXCOPYVolumes(ctx context.Context, vminfo vm.VMInfo) error {
-	openstackops := migobj.Openstackclients
-	setRDMLabel := len(vminfo.RDMDisks) > 0
-
-	for idx := range vminfo.VMDisks {
-		vmdisk := vminfo.VMDisks[idx]
-		if vmdisk.OpenstackVol == nil {
-			return fmt.Errorf("XCOPY volume for disk %s has no OpenStack volume reference", vmdisk.Name)
-		}
-		volume := vmdisk.OpenstackVol
-
-		if vminfo.UEFI {
-			if err := openstackops.SetVolumeUEFI(ctx, volume); err != nil {
-				return errors.Wrapf(err, "failed to set UEFI metadata on XCOPY volume %s", volume.ID)
-			}
-		}
-
-		if strings.ToLower(vminfo.OSType) == constants.OSFamilyWindows {
-			if err := openstackops.SetVolumeImageMetadata(ctx, volume, setRDMLabel); err != nil {
-				return errors.Wrapf(err, "failed to set Windows image metadata on XCOPY volume %s", volume.ID)
-			}
-		}
-
-		if err := openstackops.EnableQGA(ctx, volume); err != nil {
-			return errors.Wrapf(err, "failed to enable QGA on XCOPY volume %s", volume.ID)
-		}
-
-		if vmdisk.Boot {
-			if err := openstackops.SetVolumeBootable(ctx, volume); err != nil {
-				return errors.Wrapf(err, "failed to set XCOPY volume %s as bootable", volume.ID)
-			}
-		}
-	}
-
-	migobj.logMessage("Applied image metadata to XCOPY-managed volumes")
-	return nil
-}
-
-func (migobj *Migrate) AttachVolume(ctx context.Context, disk vm.VMDisk) (string, error) {
-	openstackops := migobj.Openstackclients
-	migobj.logMessage(fmt.Sprintf("Attaching volumes to VM: %s", disk.Name))
-	if disk.OpenstackVol == nil {
-		return "", errors.Wrap(fmt.Errorf("OpenStack volume is nil"), "failed to attach volume to VM")
-	}
-	volumeID := disk.OpenstackVol.ID
-	if err := openstackops.AttachVolumeToVM(ctx, volumeID); err != nil {
-		return "", errors.Wrap(err, "failed to attach volume to VM")
-	}
-
-	// Get the Path of the attached volume
-	devicePath, err := openstackops.FindDevice(volumeID)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to find device")
-	}
-	return devicePath, nil
-}
-
-func (migobj *Migrate) DetachVolume(ctx context.Context, disk vm.VMDisk) error {
-	openstackops := migobj.Openstackclients
-
-	if err := openstackops.DetachVolumeFromVM(ctx, disk.OpenstackVol.ID); err != nil {
-		return errors.Wrap(err, "failed to detach volume from VM")
-	}
-
-	err := openstackops.WaitForVolume(ctx, disk.OpenstackVol.ID)
-	if err != nil {
-		return errors.Wrap(err, "failed to wait for volume to become available")
-	}
-	return nil
-}
-
-func (migobj *Migrate) DetachAllVolumes(ctx context.Context, vminfo vm.VMInfo) error {
-	openstackops := migobj.Openstackclients
-	for _, vmdisk := range vminfo.VMDisks {
-		if vmdisk.OpenstackVol == nil {
-			migobj.logMessage(fmt.Sprintf("Skipping detach for disk %s: no OpenStack volume was created", vmdisk.Name))
-			continue
-		}
-		migobj.logMessage(fmt.Sprintf("Detaching volume %s from VM", vmdisk.Name))
-		if err := openstackops.DetachVolumeFromVM(ctx, vmdisk.OpenstackVol.ID); err != nil && !strings.Contains(err.Error(), "is not attached to volume") {
-			return errors.Wrap(err, "failed to detach volume from VM")
-		}
-
-		err := openstackops.WaitForVolume(ctx, vmdisk.OpenstackVol.ID)
-		if err != nil {
-			return errors.Wrap(err, "failed to wait for volume to become available")
-		}
-		migobj.logMessage(fmt.Sprintf("Volume %s detached from VM", vmdisk.Name))
-	}
-	time.Sleep(1 * time.Second)
-	return nil
-}
-
-// DetachAllVolumesWithCleanup is like DetachAllVolumes but handles the case where
-// a volume is attached to a foreign server (e.g. an orphaned target VM created by
-// a timed-out CreateTargetInstance call). Boot volumes cannot be hot-detached from
-// a running VM, so the foreign server is deleted first, then WaitForVolume polls
-// until Cinder reports the volume available.
-func (migobj *Migrate) DetachAllVolumesWithCleanup(ctx context.Context, vminfo vm.VMInfo) error {
-	openstackops := migobj.Openstackclients
-
-	vjailbreakUUID, err := utils.GetCurrentInstanceUUID()
-	if err != nil {
-		return errors.Wrap(err, "failed to get vJailbreak instance UUID")
-	}
-
-	deletedServers := map[string]bool{}
-
-	for _, vmdisk := range vminfo.VMDisks {
-		if vmdisk.OpenstackVol == nil {
-			migobj.logMessage(fmt.Sprintf("Skipping detach for disk %s: no OpenStack volume was created", vmdisk.Name))
-			continue
-		}
-
-		volume, err := openstackops.GetVolume(ctx, vmdisk.OpenstackVol.ID)
-		if err != nil {
-			return errors.Wrapf(err, "failed to get volume state for disk %s", vmdisk.Name)
-		}
-
-		if len(volume.Attachments) == 0 {
-			migobj.logMessage(fmt.Sprintf("Volume %s is already detached, skipping", vmdisk.Name))
-			continue
-		}
-
-		attachedServerID := volume.Attachments[0].ServerID
-
-		if attachedServerID == vjailbreakUUID {
-			migobj.logMessage(fmt.Sprintf("Detaching volume %s from vJailbreak VM", vmdisk.Name))
-			if err := openstackops.DetachVolumeFromVM(ctx, vmdisk.OpenstackVol.ID); err != nil && !strings.Contains(err.Error(), "is not attached to volume") {
-				return errors.Wrap(err, "failed to detach volume from vJailbreak VM")
-			}
-		} else {
-			if !deletedServers[attachedServerID] {
-				migobj.logMessage(fmt.Sprintf("Volume %s is attached to orphaned server %s, deleting server", vmdisk.Name, attachedServerID))
-				if err := openstackops.DeleteServer(ctx, attachedServerID); err != nil {
-					return errors.Wrapf(err, "failed to delete orphaned server %s", attachedServerID)
-				}
-				deletedServers[attachedServerID] = true
-			}
-		}
-
-		if err := openstackops.WaitForVolume(ctx, vmdisk.OpenstackVol.ID); err != nil {
-			return errors.Wrapf(err, "failed to wait for volume %s to become available", vmdisk.Name)
-		}
-		migobj.logMessage(fmt.Sprintf("Volume %s detached from VM", vmdisk.Name))
-	}
-
-	time.Sleep(1 * time.Second)
-	return nil
-}
-
-func (migobj *Migrate) verifyVMCreatedDespiteTimeout(ctx context.Context, vminfo vm.VMInfo) (string, error) {
-	vjailbreakUUID, err := utils.GetCurrentInstanceUUID()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get vJailbreak instance UUID")
-	}
-
-	var bootDisk *vm.VMDisk
-	for i := range vminfo.VMDisks {
-		if vminfo.VMDisks[i].Boot {
-			bootDisk = &vminfo.VMDisks[i]
-			break
-		}
-	}
-	if bootDisk == nil || bootDisk.OpenstackVol == nil {
-		return "", fmt.Errorf("no boot volume found")
-	}
-
-	volume, err := migobj.Openstackclients.GetVolume(ctx, bootDisk.OpenstackVol.ID)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get boot volume")
-	}
-
-	if len(volume.Attachments) == 0 {
-		return "", fmt.Errorf("boot volume is not attached to any server")
-	}
-
-	attachedServerID := volume.Attachments[0].ServerID
-	if attachedServerID == vjailbreakUUID {
-		return "", fmt.Errorf("boot volume is still attached to vJailbreak VM")
-	}
-
-	status, err := migobj.Openstackclients.GetServerStatus(ctx, attachedServerID)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get target VM status")
-	}
-
-	if strings.ToUpper(status) != "ACTIVE" {
-		return "", fmt.Errorf("target VM %s is in %s state", attachedServerID, status)
-	}
-
-	migobj.logMessage(fmt.Sprintf("Boot volume attached to active target VM %s", attachedServerID))
-	return attachedServerID, nil
-}
-
-func (migobj *Migrate) DeleteAllVolumes(ctx context.Context, vminfo vm.VMInfo) error {
-	openstackops := migobj.Openstackclients
-	for _, vmdisk := range vminfo.VMDisks {
-		if vmdisk.OpenstackVol == nil {
-			migobj.logMessage(fmt.Sprintf("Skipping delete for disk %s: no OpenStack volume was created", vmdisk.Name))
-			continue
-		}
-		err := openstackops.DeleteVolume(ctx, vmdisk.OpenstackVol.ID)
-		if err != nil {
-			return errors.Wrap(err, "failed to delete volume")
-		}
-		migobj.logMessage(fmt.Sprintf("Volume %s deleted", vmdisk.Name))
-	}
-	return nil
-}
-
-// extractFileName extracts the file name from a full VMDK path
-func extractFileName(path string) string {
-	parts := strings.Split(path, "/")
-	if len(parts) > 0 {
-		return parts[len(parts)-1]
-	}
-	return path
-}
-
-// logDiskCopyPlan logs the disk copy plan showing source to target mapping
-// Helps with debugging by showing exactly which VMDK goes to which volume
-func (migobj *Migrate) logDiskCopyPlan(vminfo vm.VMInfo) {
-	migobj.logMessage("=== Disk Copy Plan ===")
-	for idx, disk := range vminfo.VMDisks {
-		migobj.logMessage(fmt.Sprintf("[%d] %s (DeviceKey=%d): %s -> Volume %s (%s)",
-			idx,
-			disk.Name,
-			disk.Disk.Key,
-			extractFileName(disk.SnapBackingDisk),
-			disk.OpenstackVol.Name,
-			disk.Path))
-	}
-}
-
-// validateDiskMapping validates that disk mapping is correct before starting copy
-// Cross-checks vminfo data with nbdops to ensure correct source-to-target mapping
-func (migobj *Migrate) validateDiskMapping(vminfo vm.VMInfo) error {
-	migobj.logMessage("Validating disk mapping before copy operation...")
-
-	// Verify number of disks matches number of NBD servers
-	if len(vminfo.VMDisks) != len(migobj.Nbdops) {
-		return fmt.Errorf("disk count mismatch: vminfo has %d disks but %d NBD servers configured", len(vminfo.VMDisks), len(migobj.Nbdops))
-	}
-
-	for idx, vmdisk := range vminfo.VMDisks {
-		// Validate volume exists
-		if vmdisk.OpenstackVol == nil {
-			return fmt.Errorf("OpenStack volume is nil for disk %s (DeviceKey=%d)", vmdisk.Name, vmdisk.Disk.Key)
-		}
-
-		// Validate device path exists
-		if vmdisk.Path == "" {
-			return fmt.Errorf("device path is empty for disk %s (DeviceKey=%d)", vmdisk.Name, vmdisk.Disk.Key)
-		}
-
-		// Validate snapshot backing disk exists
-		if vmdisk.SnapBackingDisk == "" {
-			return fmt.Errorf("snapshot backing disk is empty for disk %s (DeviceKey=%d)", vmdisk.Name, vmdisk.Disk.Key)
-		}
-
-		// Cross-check: verify NBD server at this index is initialized
-		if migobj.Nbdops[idx] == nil {
-			return fmt.Errorf("NBD server not initialized for disk %d (%s)", idx, vmdisk.Name)
-		}
-
-		// Log validation details for this disk
-		utils.PrintLog(fmt.Sprintf("[%d] Validated %s (DeviceKey=%d): SnapFile=%s, Volume=%s, Path=%s",
-			idx, vmdisk.Name, vmdisk.Disk.Key,
-			extractFileName(vmdisk.SnapBackingDisk),
-			vmdisk.OpenstackVol.ID, vmdisk.Path))
-	}
-
-	migobj.logMessage("Disk mapping validation passed")
-	return nil
-}
-
-// This function enables CBT on the VM if it is not enabled and takes a snapshot for initializing CBT
-func (migobj *Migrate) EnableCBTWrapper() error {
-	vmops := migobj.VMops
-
-	// If it is cold migration we do not need to enable CBT.
-	if migobj.MigrationType == "cold" {
-		migobj.logMessage("Cold migration selected: skipping CBT check and enablement")
-		return nil
-	}
-
-	// CBT requires virtual hardware >= 7. Fail with a clear error if
-	// changeTrackingEnabled is unavailable, instead of panicking.
-	hwVersion, VersionErr := vmops.GetHardwareVersion()
-	if VersionErr != nil {
-		migobj.logMessage(fmt.Sprintf("Could not determine hardware version, Trying to enable CBT: %s", VersionErr))
-	}
-	migobj.logMessage(fmt.Sprintf("Hardware version detected: %s", hwVersion))
-
-	if hwVersion > 0 && hwVersion < vm.MinCBTHardwareVersion {
-		return fmt.Errorf(
-			"changed block tracking (CBT) is not supported on virtual hardware version %d "+
-				"(requires version %d or newer); please use cold migration for this VM",
-			hwVersion, vm.MinCBTHardwareVersion)
-	}
-
-	cbt, err := vmops.IsCBTEnabled()
-	if err != nil {
-		return errors.Wrap(err, "failed to check if CBT is enabled")
-	}
-	migobj.logMessage(fmt.Sprintf("CBT Enabled: %t", cbt))
-
-	if !cbt {
-		// 7.5. Enable CBT
-		migobj.logMessage("CBT is not enabled. Enabling CBT")
-		err = vmops.EnableCBT()
-		if err != nil {
-			return errors.Wrap(err, "failed to enable CBT")
-		}
-		_, err := vmops.IsCBTEnabled()
-		if err != nil {
-			return errors.Wrap(err, "failed to check if CBT is enabled")
-		}
-		migobj.logMessage("Creating temporary snapshot of the source VM")
-		err = vmops.TakeSnapshot("tmp-snap")
-		if err != nil {
-			return errors.Wrap(err, "failed to take snapshot of source VM")
-		}
-		utils.PrintLog("Snapshot created successfully")
-		err = vmops.DeleteSnapshot("tmp-snap")
-		if err != nil {
-			return errors.Wrap(err, "failed to delete snapshot of source VM")
-		}
-		utils.PrintLog("Snapshot deleted successfully")
-		migobj.logMessage("CBT enabled successfully")
-	}
-	return nil
 }
 
 func (migobj *Migrate) WaitforCutover() error {
@@ -557,264 +135,6 @@ func (migobj *Migrate) WaitforCutover() error {
 		}
 	}
 	return nil
-}
-func (migobj *Migrate) SyncCBT(ctx context.Context, vminfo vm.VMInfo) error {
-	migobj.logMessage("Starting Periodic sync process")
-	defer migobj.logMessage("Periodic sync process completed")
-	vmops := migobj.VMops
-	nbdops := migobj.Nbdops
-	envURL := migobj.URL
-	envUserName := migobj.UserName
-	envPassword := migobj.Password
-	thumbprint := migobj.Thumbprint
-	migration_snapshot, err := vmops.GetSnapshot(constants.MigrationSnapshotName)
-	if err != nil {
-		return errors.Wrap(err, "failed to get snapshot")
-	}
-
-	var changedAreas types.DiskChangeInfo
-
-	for idx := range vminfo.VMDisks {
-		changedAreas, err = vmops.CustomQueryChangedDiskAreas(vminfo.VMDisks[idx].ChangeID, migration_snapshot, vminfo.VMDisks[idx].Disk, 0)
-		if err != nil {
-			return errors.Wrap(err, "failed to get changed disk areas")
-		}
-
-		if len(changedAreas.ChangedArea) == 0 {
-			migobj.logMessage(fmt.Sprintf("Periodic Sync: Disk %d: No changed blocks found. Skipping copy", idx))
-		} else {
-			migobj.logMessage(fmt.Sprintf("Periodic Sync: Disk %d: Blocks have Changed.", idx))
-
-			// Before starting NBD server, update disk info with new snapshot details
-			// We have marked block copy as false, in order to not update changeID.
-			// This should now update the snapname and snapBackingDisk with the new snapshot details and copy correctly.
-			err = vmops.UpdateDiskInfo(&vminfo, vminfo.VMDisks[idx], false)
-			if err != nil {
-				return errors.Wrap(err, "failed to update disk info")
-			}
-
-			utils.PrintLog("Restarting NBD server")
-			err = nbdops[idx].StopNBDServer()
-			if err != nil {
-				return errors.Wrap(err, "failed to stop NBD server")
-			}
-
-			err = nbdops[idx].StartNBDServer(vmops.GetVMObj(), envURL, envUserName, envPassword, thumbprint, vminfo.VMDisks[idx].Snapname, vminfo.VMDisks[idx].SnapBackingDisk, migobj.EventReporter)
-			if err != nil {
-				return errors.Wrap(err, "failed to start NBD server")
-			}
-			// sleep for 2 seconds to allow the NBD server to start
-			time.Sleep(2 * time.Second)
-
-			// 11. Copy Changed Blocks over
-			changedBlockCopySuccess := true
-			startTime := time.Now()
-			migobj.logMessage(fmt.Sprintf("Periodic Sync: Starting incremental block copy for disk %d at %s", idx, startTime))
-			err = nbdops[idx].CopyChangedBlocks(ctx, changedAreas, vminfo.VMDisks[idx].Path, vminfo.VMDisks[idx].OpenstackVol.Encrypted)
-			if err != nil {
-				migobj.logMessage(fmt.Sprintf("Periodic Sync: Failed to copy changed blocks for disk %d: %v", idx, err))
-				select {
-				case <-ctx.Done():
-					err = vmops.CleanUpSnapshots(false)
-					changedBlockCopySuccess = false
-					if err != nil {
-						return errors.Wrap(err, "failed to cleanup snapshot of source VM")
-					}
-				default:
-					return errors.Wrap(err, "failed to copy changed blocks")
-				}
-			}
-
-			duration := time.Since(startTime)
-
-			migobj.logMessage(fmt.Sprintf("Periodic Sync: Incremental block copy for disk %d completed in %s", idx, duration))
-
-			err = vmops.UpdateDiskInfo(&vminfo, vminfo.VMDisks[idx], changedBlockCopySuccess)
-			if err != nil {
-				return errors.Wrap(err, "failed to update disk info")
-			}
-			if !changedBlockCopySuccess {
-				migobj.logMessage(fmt.Sprintf("Periodic Sync: Failed to copy changed blocks: %s", err))
-				migobj.logMessage(fmt.Sprintf("Periodic Sync: Since full copy has completed, Retrying copy of changed blocks for disk: %d", idx))
-			}
-		}
-	}
-	// Cleanup the snapshot taken for incremental copy
-	return nil
-}
-func (migobj *Migrate) getSyncEnabled() bool {
-	var enabled bool
-	enabled = false
-	migrationParams, err := utils.GetMigrationParams(context.Background(), migobj.K8sClient)
-	if err != nil {
-		return enabled
-	}
-	if migrationParams.PeriodicSyncEnabled {
-		enabled = true
-	}
-	return enabled
-}
-func (migobj *Migrate) getSyncDuration() time.Duration {
-	const defaultInterval = "1h"
-
-	migobj.logMessage("Periodic Sync: Setting up sync interval")
-
-	migrationParams, err := utils.GetMigrationParams(context.Background(), migobj.K8sClient)
-	if err != nil {
-		migobj.logMessage(fmt.Sprintf("WARNING: Failed to get migration params: %v, using default interval (%s)",
-			err, defaultInterval))
-	}
-	// Get sync interval settings
-	interval := migrationParams.PeriodicSyncInterval
-	if interval == "" {
-		vjailbreakSettings, err := k8sutils.GetVjailbreakSettings(context.Background(), migobj.K8sClient)
-		if err != nil {
-			migobj.logMessage(fmt.Sprintf("WARNING: Failed to get vjailbreak settings: %v, using default interval (%s)",
-				err, defaultInterval))
-		}
-		interval = vjailbreakSettings.PeriodicSyncInterval
-		if interval == "" {
-			interval = defaultInterval
-		}
-	}
-	// Calculate wait time based on unit
-	waitTime, err := time.ParseDuration(interval)
-	if err != nil {
-		migobj.logMessage(fmt.Sprintf("WARNING: Failed to parse interval %s, using default interval (%s)", interval, defaultInterval))
-		interval = defaultInterval
-		waitTime, _ = time.ParseDuration(interval)
-	} else if waitTime < 5*time.Minute {
-		migobj.logMessage(fmt.Sprintf("WARNING: Interval %s is less than 5 minutes, falling back to 5m", interval))
-		waitTime = 5 * time.Minute
-	}
-	return waitTime
-}
-
-func (migobj *Migrate) WaitforAdminCutover(ctx context.Context, vminfo vm.VMInfo) error {
-	vmops := migobj.VMops
-	maxRetries, capInterval := utils.GetRetryLimits()
-	migobj.logMessage(constants.EventMessageWaitingForAdminCutOver)
-
-	// Initialize state machine context
-	syncCtx := &PeriodicSyncContext{
-		CurrentState: StateIdle,
-	}
-
-	elapsed := time.Duration(0)
-
-	for {
-		syncEnabled := migobj.getSyncEnabled()
-		var syncInterval time.Duration
-		if syncEnabled {
-			syncInterval = migobj.getSyncDuration()
-		}
-
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case label := <-migobj.PodLabelWatcher:
-			if label == "yes" && syncCtx.CurrentState == StateIdle {
-				migobj.logMessage("Admin cutover triggered")
-				return nil
-			}
-		default:
-			if !syncEnabled {
-				// Small sleep to prevent busy-waiting when sync is disabled
-				time.Sleep(1 * time.Second)
-				continue
-			}
-
-			// Calculate wait time based on elapsed time
-			if elapsed >= syncInterval {
-				migobj.logMessage("Periodic Sync: Previous sync took longer than interval, starting next cycle immediately")
-				elapsed = syncInterval
-			}
-
-			waitTime := syncInterval - elapsed
-			stateInfo := syncCtx.CurrentState.String()
-			if syncCtx.WarningMessage != "" {
-				stateInfo = fmt.Sprintf("%s (WARNING: %s)", stateInfo, syncCtx.WarningMessage)
-			}
-			migobj.logMessage(fmt.Sprintf("Periodic Sync: Waiting %s before next sync cycle (state: %s)", waitTime, stateInfo))
-
-			select {
-			case <-ctx.Done():
-				return ctx.Err()
-			case label := <-migobj.PodLabelWatcher:
-				if label == "yes" {
-					migobj.logMessage("Admin cutover triggered during wait")
-					return nil
-				}
-			case <-time.After(waitTime):
-				// wait completed → proceed to sync
-			}
-
-			// Execute state machine - always start fresh from Idle each cycle
-			migobj.logMessage(fmt.Sprintf("Periodic Sync: Starting sync cycle (interval: %s)", syncInterval))
-			start := time.Now()
-
-			// Reset state to start fresh each cycle
-			syncCtx.CurrentState = StateCleaningSnapshots
-
-			if err := migobj.Vcclient.EnsureSessionActive(ctx); err != nil {
-				syncCtx.LastError = err
-				syncCtx.WarningMessage = fmt.Sprintf("vCenter session refresh failed: %v. Will retry on next sync interval.", err)
-				syncCtx.CurrentState = StateIdle
-				migobj.logMessage(fmt.Sprintf("Periodic Sync: WARNING - %s", syncCtx.WarningMessage))
-				elapsed = time.Since(start)
-				continue
-			}
-
-			// State: CleaningSnapshots
-			err := utils.DoRetryWithExponentialBackoff(ctx, func() error {
-				return vmops.CleanUpSnapshots(false)
-			}, maxRetries, capInterval)
-			if err != nil {
-				syncCtx.LastError = err
-				syncCtx.WarningMessage = fmt.Sprintf("Snapshot cleanup failed after %d retries: %v. Will retry on next sync interval.", maxRetries, err)
-				syncCtx.CurrentState = StateIdle
-				migobj.logMessage(fmt.Sprintf("Periodic Sync: WARNING - %s", syncCtx.WarningMessage))
-				elapsed = time.Since(start)
-				continue
-			}
-
-			// State: TakingSnapshot
-			syncCtx.CurrentState = StateTakingSnapshot
-			err = utils.DoRetryWithExponentialBackoff(ctx, func() error {
-				return vmops.TakeSnapshot(constants.MigrationSnapshotName)
-			}, maxRetries, capInterval)
-			if err != nil {
-				syncCtx.LastError = err
-				syncCtx.WarningMessage = fmt.Sprintf("Snapshot creation '%s' failed after %d retries: %v. Will retry on next sync interval.", constants.MigrationSnapshotName, maxRetries, err)
-				syncCtx.CurrentState = StateIdle
-				migobj.logMessage(fmt.Sprintf("Periodic Sync: WARNING - %s", syncCtx.WarningMessage))
-				elapsed = time.Since(start)
-				continue
-			}
-
-			// State: SyncingCBT
-			syncCtx.CurrentState = StateSyncingCBT
-			err = utils.DoRetryWithExponentialBackoff(ctx, func() error {
-				return migobj.SyncCBT(ctx, vminfo)
-			}, maxRetries, capInterval)
-			if err != nil {
-				syncCtx.LastError = err
-				syncCtx.WarningMessage = fmt.Sprintf("CBT sync failed after %d retries: %v. Will retry on next sync interval.", maxRetries, err)
-				syncCtx.CurrentState = StateIdle
-				migobj.logMessage(fmt.Sprintf("Periodic Sync: WARNING - %s", syncCtx.WarningMessage))
-				elapsed = time.Since(start)
-				continue
-			}
-
-			// Sync completed successfully - clear warning state
-			syncCtx.CurrentState = StateIdle
-			syncCtx.WarningMessage = ""
-			syncCtx.LastError = nil
-			migobj.logMessage("Periodic Sync: Sync cycle completed successfully")
-
-			elapsed = time.Since(start)
-		}
-	}
 }
 
 func (migobj *Migrate) CheckIfAdminCutoverSelected() bool {
@@ -832,6 +152,7 @@ func (migobj *Migrate) CheckIfAdminCutoverSelected() bool {
 	}
 	return false
 }
+
 func (migobj *Migrate) CheckCutoverOptions() (bool, string) {
 	if migobj.Reporter == nil {
 		return false, ""
@@ -846,1226 +167,6 @@ func (migobj *Migrate) CheckCutoverOptions() (bool, string) {
 		return true, value
 	}
 	return false, ""
-}
-
-func (migobj *Migrate) LiveReplicateDisks(ctx context.Context, vminfo vm.VMInfo) (vm.VMInfo, error) {
-	vmops := migobj.VMops
-	nbdops := migobj.Nbdops
-	envURL := migobj.URL
-	envUserName := migobj.UserName
-	envPassword := migobj.Password
-	thumbprint := migobj.Thumbprint
-
-	// Get migration parameters to check if user acknowledged network conflict risk
-	migrationParams, err := utils.GetMigrationParams(ctx, migobj.K8sClient)
-	if err != nil {
-		migobj.logMessage(fmt.Sprintf("WARNING: Failed to get migration params: %v, continuing with migration", err))
-	} else {
-		if migobj.MigrationType == "mock" {
-
-			if migrationParams.AcknowledgeNetworkConflictRisk {
-				migobj.logMessage("User acknowledged the risk involved")
-			} else {
-				migobj.logMessage("User did not acknowledge the risk involved")
-			}
-		}
-	}
-
-	cutoverLabelPresent, cutoverLabelValue := migobj.CheckCutoverOptions()
-	// if the cutover immediately is selected with cold migration type then the migration will happen like cold migration
-	var currentCutoverOption string
-	if migobj.MigrationType == "cold" {
-		if cutoverLabelValue != "" {
-			if cutoverLabelValue == "yes" {
-
-				currentCutoverOption = "Immediately After Data Copy"
-			} else if cutoverLabelValue == "no" {
-				currentCutoverOption = "Admin Initiated Cutover"
-			}
-			migobj.logMessage(fmt.Sprintf("Migration Type : %s | Cutover Option %s", migobj.MigrationType, currentCutoverOption))
-		}
-		if err := vmops.VMPowerOff(); err != nil {
-			return vminfo, errors.Wrap(err, "failed to power off VM")
-		}
-		// Verify VM is actually powered off
-		if err := utils.DoRetryWithExponentialBackoff(ctx, func() error {
-			currState, stateErr := vmops.GetVMObj().PowerState(ctx)
-			if stateErr != nil {
-				return stateErr
-			}
-			if currState != types.VirtualMachinePowerStatePoweredOff {
-				return fmt.Errorf("VM power-off command completed but VM is still in state: %s", currState)
-			}
-			return nil
-		}, constants.MaxPowerOffRetryLimit, constants.PowerOffRetryCap); err != nil {
-			return vminfo, errors.Wrap(err, "failed to verify VM power state after power off")
-		}
-	}
-
-	// clean up snapshots
-	utils.PrintLog("Cleaning up snapshots before copy")
-	err = vmops.CleanUpSnapshots(false)
-	if err != nil {
-		return vminfo, errors.Wrap(err, "failed to clean up snapshots: %s, please delete manually before starting again")
-	}
-
-	utils.PrintLog("Starting NBD server")
-	err = vmops.TakeSnapshot(constants.MigrationSnapshotName)
-	if err != nil {
-		return vminfo, errors.Wrap(err, "failed to take snapshot of source VM")
-	}
-
-	// Cold migrations never use CBT, so the snapshot disks have no change IDs
-	// (especially on legacy hardware version < 7). Only require change IDs for
-	// non-cold migrations, which rely on them for incremental copies.
-	err = vmops.UpdateDisksInfo(&vminfo, migobj.MigrationType != "cold")
-	if err != nil {
-		return vminfo, errors.Wrap(err, "failed to update disk info")
-	}
-
-	for idx, vmdisk := range vminfo.VMDisks {
-		migobj.logMessage(fmt.Sprintf("Copying disk %d, Completed: 0%%", idx))
-		err = nbdops[idx].StartNBDServer(vmops.GetVMObj(), envURL, envUserName, envPassword, thumbprint, vmdisk.Snapname, vmdisk.SnapBackingDisk, migobj.EventReporter)
-		if err != nil {
-			return vminfo, errors.Wrap(err, "failed to start NBD server")
-		}
-	}
-	// sleep for 2 seconds to allow the NBD server to start
-	time.Sleep(2 * time.Second)
-	final := false
-
-	for idx, vmdisk := range vminfo.VMDisks {
-		vminfo.VMDisks[idx].Path, err = migobj.AttachVolume(ctx, vmdisk)
-		if err != nil {
-			return vminfo, errors.Wrap(err, "failed to attach volume")
-		}
-	}
-
-	// Validate disk mapping before starting copy
-	if err := migobj.validateDiskMapping(vminfo); err != nil {
-		return vminfo, errors.Wrap(err, "disk mapping validation failed")
-	}
-
-	// Log the disk copy plan for debugging
-	migobj.logDiskCopyPlan(vminfo)
-
-	vcenterSettings, err := k8sutils.GetVjailbreakSettings(ctx, migobj.K8sClient)
-	if err != nil {
-		return vminfo, errors.Wrap(err, "failed to get vcenter settings")
-	}
-	utils.PrintLog(fmt.Sprintf("Fetched vjailbreak settings for Changed Blocks Copy Iteration Threshold: %d", vcenterSettings.ChangedBlocksCopyIterationThreshold))
-
-	// Check if migration has admin cutover if so don't copy any more changed blocks
-	adminInitiatedCutover := cutoverLabelPresent && (cutoverLabelValue == "no")
-	incrementalCopyCount := 0
-	for {
-		// If its the first copy, copy the entire disk
-		if incrementalCopyCount == 0 {
-			for idx := range vminfo.VMDisks {
-				startTime := time.Now()
-				disk := vminfo.VMDisks[idx]
-
-				migobj.logMessage(fmt.Sprintf("Starting full disk copy [%d/%d]: %s (DeviceKey=%d)",
-					idx+1, len(vminfo.VMDisks), disk.Name, disk.Disk.Key))
-				migobj.logMessage(fmt.Sprintf("  Source: %s", extractFileName(disk.SnapBackingDisk)))
-				migobj.logMessage(fmt.Sprintf("  Target: %s (Volume ID: %s)", disk.Path, disk.OpenstackVol.ID))
-
-				err = nbdops[idx].CopyDisk(ctx, disk.Path, idx, disk.OpenstackVol.Encrypted)
-				if err != nil {
-					return vminfo, errors.Wrap(err, fmt.Sprintf("failed to copy disk %s (DeviceKey=%d)", disk.Name, disk.Disk.Key))
-				}
-				duration := time.Since(startTime)
-				if migobj.MigrationType == "cold" {
-					migobj.logMessage(fmt.Sprintf("✓ Disk %d (%s) copied successfully in %s", idx, disk.Name, duration))
-				} else {
-					migobj.logMessage(fmt.Sprintf("✓ Disk %d (%s) copied successfully in %s, copying changed blocks now", idx, disk.Name, duration))
-				}
-			}
-
-			if adminInitiatedCutover {
-				utils.PrintLog("Admin initiated cutover detected, skipping changed blocks copy")
-				if err := migobj.WaitforAdminCutover(ctx, vminfo); err != nil {
-					return vminfo, errors.Wrap(err, "failed to start VM Cutover")
-				}
-				if err := migobj.Vcclient.EnsureSessionActive(ctx); err != nil {
-					return vminfo, errors.Wrap(err, "failed to refresh vCenter session post-cutover")
-				}
-				if migobj.MigrationType == "mock" {
-					utils.PrintLog("Mock migration detected, skipping VM power off")
-				} else {
-					utils.PrintLog("Shutting down source VM and performing final copy")
-					err = vmops.VMPowerOff()
-					if err != nil {
-						return vminfo, errors.Wrap(err, "failed to power off VM")
-					}
-					// Verify VM is actually powered off
-					if err := utils.DoRetryWithExponentialBackoff(ctx, func() error {
-						currState, stateErr := vmops.GetVMObj().PowerState(ctx)
-						if stateErr != nil {
-							return stateErr
-						}
-						if currState != types.VirtualMachinePowerStatePoweredOff {
-							return fmt.Errorf("VM power-off command completed but VM is still in state: %s", currState)
-						}
-						return nil
-					}, constants.MaxPowerOffRetryLimit, constants.PowerOffRetryCap); err != nil {
-						return vminfo, errors.Wrap(err, "failed to verify VM power state after power off")
-					}
-				}
-			}
-			if err := migobj.WaitforCutover(); err != nil {
-				return vminfo, errors.Wrap(err, "failed to start VM Cutover")
-			}
-		} else {
-			migration_snapshot, err := vmops.GetSnapshot(constants.MigrationSnapshotName)
-			if err != nil {
-				return vminfo, errors.Wrap(err, "failed to get snapshot")
-			}
-
-			var changedAreas types.DiskChangeInfo
-			done := true
-
-			for idx := range vminfo.VMDisks {
-				err := vmops.UpdateDiskInfo(&vminfo, vminfo.VMDisks[idx], false)
-				if err != nil {
-					return vminfo, errors.Wrap(err, "failed to update disk info")
-				}
-
-				changedAreas, err = vmops.CustomQueryChangedDiskAreas(vminfo.VMDisks[idx].ChangeID, migration_snapshot, vminfo.VMDisks[idx].Disk, 0)
-				if err != nil {
-					return vminfo, errors.Wrap(err, "failed to get changed disk areas")
-				}
-
-				if len(changedAreas.ChangedArea) == 0 {
-					if migobj.MigrationType != "cold" {
-						migobj.logMessage(fmt.Sprintf("Disk %d: No changed blocks found. Skipping copy", idx))
-					}
-				} else {
-					migobj.logMessage(fmt.Sprintf("Disk %d: Blocks have Changed.", idx))
-
-					utils.PrintLog("Restarting NBD server")
-					err = nbdops[idx].StopNBDServer()
-					if err != nil {
-						return vminfo, errors.Wrap(err, "failed to stop NBD server")
-					}
-
-					err = nbdops[idx].StartNBDServer(vmops.GetVMObj(), envURL, envUserName, envPassword, thumbprint, vminfo.VMDisks[idx].Snapname, vminfo.VMDisks[idx].SnapBackingDisk, migobj.EventReporter)
-					if err != nil {
-						return vminfo, errors.Wrap(err, "failed to start NBD server")
-					}
-					// sleep for 2 seconds to allow the NBD server to start
-					time.Sleep(2 * time.Second)
-
-					// 11. Copy Changed Blocks over
-					done = false
-					changedBlockCopySuccess := true
-					migobj.logMessage("Copying changed blocks")
-
-					startTime := time.Now()
-					migobj.logMessage(fmt.Sprintf("Starting incremental block copy for disk %d at %s", idx, startTime))
-
-					// Use exponential backoff for retry logic (3 retries, 30 second cap)
-					copyErr := utils.DoRetryWithExponentialBackoff(ctx, func() error {
-						return nbdops[idx].CopyChangedBlocks(ctx, changedAreas, vminfo.VMDisks[idx].Path, vminfo.VMDisks[idx].OpenstackVol.Encrypted)
-					}, 3, 30*time.Second)
-
-					if copyErr != nil {
-						changedBlockCopySuccess = false
-						// Fail the migration if copy fails after 3 retries
-						return vminfo, errors.Wrap(copyErr, fmt.Sprintf("failed to copy changed blocks for disk %d after 3 attempts", idx))
-					}
-
-					duration := time.Since(startTime)
-
-					migobj.logMessage(fmt.Sprintf("Incremental block copy for disk %d completed in %s", idx, duration))
-
-					err = vmops.UpdateDiskInfo(&vminfo, vminfo.VMDisks[idx], changedBlockCopySuccess)
-					if err != nil {
-						return vminfo, errors.Wrap(err, "failed to update disk info")
-					}
-					migobj.logMessage(fmt.Sprintf("Finished copying and syncing changed blocks for disk %d in %s [Progress: %d/20]", idx, duration, incrementalCopyCount))
-				}
-			}
-			if final {
-				break
-			}
-			if done || incrementalCopyCount > vcenterSettings.ChangedBlocksCopyIterationThreshold {
-				if migobj.MigrationType == "mock" {
-					utils.PrintLog("Mock migration detected, skipping VM power off")
-				} else {
-					utils.PrintLog("Shutting down source VM and performing final copy")
-					err = vmops.VMPowerOff()
-					if err != nil {
-						return vminfo, errors.Wrap(err, "failed to power off VM")
-					}
-					// Verify VM is actually powered off
-					if err := utils.DoRetryWithExponentialBackoff(ctx, func() error {
-						currState, stateErr := vmops.GetVMObj().PowerState(ctx)
-						if stateErr != nil {
-							return stateErr
-						}
-						if currState != types.VirtualMachinePowerStatePoweredOff {
-							return fmt.Errorf("VM power-off command completed but VM is still in state: %s", currState)
-						}
-						return nil
-					}, constants.MaxPowerOffRetryLimit, constants.PowerOffRetryCap); err != nil {
-						return vminfo, errors.Wrap(err, "failed to verify VM power state after power off")
-					}
-				}
-				final = true
-			}
-		}
-
-		// Update old change id to the new base change id value
-		// Only do this after you have gone through all disks with old change id.
-		// If you dont, only your first disk will have the updated changes
-
-		err = vmops.CleanUpSnapshots(false)
-		if err != nil {
-			return vminfo, errors.Wrap(err, "failed to cleanup snapshot of source VM")
-		}
-
-		// If migration type is cold, check power state and exit the live replicate function.
-		if migobj.MigrationType == "cold" {
-			// Verify VM is actually powered off
-			if err := utils.DoRetryWithExponentialBackoff(ctx, func() error {
-				currState, stateErr := vmops.GetVMObj().PowerState(ctx)
-				if stateErr != nil {
-					return stateErr
-				}
-				if currState != types.VirtualMachinePowerStatePoweredOff {
-					return fmt.Errorf("Cold migration requires the VM to remain powered off. However, the VM was found in the '%s' state after power-off and snapshot copy. Aborting migration to avoid data inconsistency.", currState)
-				}
-				return nil
-			}, constants.MaxPowerOffRetryLimit, constants.PowerOffRetryCap); err != nil {
-				return vminfo, errors.Wrap(err, "failed to verify VM power state after power off")
-			}
-			break
-		}
-
-		err = vmops.TakeSnapshot(constants.MigrationSnapshotName)
-		if err != nil {
-			return vminfo, errors.Wrap(err, "failed to take snapshot of source VM")
-		}
-
-		incrementalCopyCount += 1
-
-	}
-
-	err = migobj.DetachAllVolumes(ctx, vminfo)
-	if err != nil {
-		return vminfo, errors.Wrap(err, "Failed to detach all volumes from VM")
-	}
-
-	utils.PrintLog("Stopping NBD server")
-	for _, nbdserver := range nbdops {
-		err = nbdserver.StopNBDServer()
-		if err != nil {
-			return vminfo, errors.Wrap(err, "failed to stop NBD server")
-		}
-	}
-
-	utils.PrintLog("Deleting migration snapshot")
-	err = vmops.CleanUpSnapshots(true)
-	if err != nil {
-		migobj.logMessage(fmt.Sprintf(`Failed to cleanup snapshot of source VM: %s, since copy is completed, 
-        continuing with the migration`, err))
-	}
-	return vminfo, nil
-}
-
-// getBootCommand returns the appropriate command to detect boot volume based on OS type
-func (migobj *Migrate) getBootCommand(osType string) string {
-	switch strings.ToLower(osType) {
-	case constants.OSFamilyWindows:
-		return "ls /Windows"
-	case constants.OSFamilyLinux:
-		return "ls /boot"
-	default:
-		return "inspect-os"
-	}
-}
-
-// attachAllVolumes attaches all volumes and updates their paths in vminfo
-func (migobj *Migrate) attachAllVolumes(ctx context.Context, vminfo *vm.VMInfo) error {
-	for idx, vmdisk := range vminfo.VMDisks {
-		path, err := migobj.AttachVolume(ctx, vmdisk)
-		if err != nil {
-			return errors.Wrap(err, "failed to attach volume")
-		}
-		vminfo.VMDisks[idx].Path = path
-	}
-	return nil
-}
-
-// detectBootVolume identifies which volume contains the boot partition
-func (migobj *Migrate) detectBootVolume(vminfo vm.VMInfo, getBootCommand string) (bootVolumeIndex int, osPath string, err error) {
-	bootVolumeIndex = -1
-
-	utils.PrintLog(fmt.Sprintf("Detecting boot volume (UEFI: %t)", vminfo.UEFI))
-
-	for idx := range vminfo.VMDisks {
-		ans, cmdErr := virtv2v.RunCommandInGuest(vminfo.VMDisks[idx].Path, getBootCommand, false)
-		if cmdErr != nil || ans == "" {
-			continue
-		}
-
-		utils.PrintLog(fmt.Sprintf("Boot volume detected: Disk %d (%s)", idx, vminfo.VMDisks[idx].Name))
-		osPath = strings.TrimSpace(ans)
-		bootVolumeIndex = idx
-		break
-	}
-
-	if bootVolumeIndex < 0 {
-		utils.PrintLog("WARNING: No boot volume detected")
-	}
-
-	return bootVolumeIndex, osPath, nil
-}
-
-// handleLinuxOSDetection handles OS detection and validation for Linux systems
-func (migobj *Migrate) handleLinuxOSDetection(vminfo vm.VMInfo, bootVolumeIndex int, osPath string, autoFstabUpdate bool) (finalBootIndex int, finalOsPath string, osRelease string, espDiskIndex int, err error) {
-	finalBootIndex = bootVolumeIndex
-	finalOsPath = osPath
-
-	// Nothing downstream is safe to index if there are no disks.
-	if len(vminfo.VMDisks) == 0 {
-		return -1, "", "", -1, errors.New("no disks present for VM; cannot detect boot disk")
-	}
-
-	// Run get-bootable-partition.sh script
-	var ans string
-	var cmdErr error
-
-	if ans, cmdErr = virtv2v.RunGetBootablePartitionScript(vminfo.VMDisks); cmdErr != nil {
-		utils.PrintLog(fmt.Sprintf("Warning: Failed to run get-bootable-partition.sh: %v", cmdErr))
-	} else if ans != "" {
-		migobj.logMessage(fmt.Sprintf("Bootable partition: %s", ans))
-	}
-
-	if ans == "" {
-		return -1, "", "", -1, errors.New("empty bootable partition from the script")
-	}
-
-	index, err := virtv2v.RunCommandInGuestAllVolumes(vminfo.VMDisks, "device-index", false, strings.TrimSpace(ans))
-	if err != nil {
-		fmt.Printf("failed to run command (%s): %v: %s\n", index, err, strings.TrimSpace(index))
-		return -1, "", "", -1, err
-	}
-
-	finalBootIndex, err = strconv.Atoi(strings.TrimSpace(index))
-	if err != nil {
-		return -1, "", "", -1, errors.Wrap(err, "failed to convert bootable partition index to int")
-	}
-	migobj.logMessage(fmt.Sprintf("Bootable partition index: %d", finalBootIndex))
-
-	// device-index can resolve to the libguestfs appliance's own disk (a slot past
-	// the guest's disks). Fail the migration rather than indexing vminfo.VMDisks out
-	// of range (which would panic) or silently converting the wrong disk.
-	if finalBootIndex < 0 || finalBootIndex >= len(vminfo.VMDisks) {
-		return -1, "", "", -1, errors.Errorf(
-			"detected boot disk index %d is out of range for %d disk(s); aborting migration",
-			finalBootIndex, len(vminfo.VMDisks))
-	}
-
-	// Detect ESP for UEFI VMs
-	espDiskIndex = -1
-	if vminfo.UEFI {
-		detectedESPIndex, espErr := virtv2v.DetectESPDiskIndex(vminfo.VMDisks)
-		if espErr != nil {
-			migobj.logMessage(fmt.Sprintf("Error detecting ESP disk: %v", espErr))
-		} else if detectedESPIndex >= 0 && detectedESPIndex < len(vminfo.VMDisks) {
-			espDiskIndex = detectedESPIndex
-			migobj.logMessage(fmt.Sprintf("ESP detected on Disk %d: %s", espDiskIndex, vminfo.VMDisks[espDiskIndex].Name))
-		} else {
-			migobj.logMessage("WARNING: No ESP detected for UEFI VM")
-		}
-	}
-
-	// Use boot partition device as OS path for virt-v2v
-	// virt-v2v-in-place will auto-detect the actual OS location (LVM or regular partition)
-	// when all disks are provided in the libvirt XML
-	finalOsPath = strings.TrimSpace(ans)
-	migobj.logMessage(fmt.Sprintf("OS path for virt-v2v: %s", finalOsPath))
-
-	osRelease, err = virtv2v.GetOsReleaseAllVolumes(vminfo.VMDisks)
-	if err != nil {
-		return -1, "", "", -1, errors.Wrapf(err, "failed to get os release: %s", strings.TrimSpace(osRelease))
-	}
-
-	if err := migobj.validateLinuxOS(osRelease); err != nil {
-		return -1, "", "", -1, err
-	}
-
-	// Run generate-mount-persistence.sh script based on AUTO_FSTAB_UPDATE setting.
-	// The flag passed to the script varies by OS: SUSE GRUB Legacy guests use
-	// --replace-fstab to avoid rewriting device.map before virt-v2v runs (see
-	// RunMountPersistenceScript for the full rationale).
-	if autoFstabUpdate {
-		migobj.logMessage("Running generate-mount-persistence.sh script")
-		if err := virtv2v.RunMountPersistenceScript(vminfo.VMDisks, vminfo.VMDisks[finalBootIndex].Path, osRelease); err != nil {
-			utils.PrintLog(fmt.Sprintf("Warning: Failed to run generate-mount-persistence.sh: %v", err))
-			// Don't fail the migration, just log the warning
-		} else {
-			migobj.logMessage("Successfully ran generate-mount-persistence.sh script")
-		}
-	} else {
-		migobj.logMessage("Skipping generate-mount-persistence.sh script (AUTO_FSTAB_UPDATE is disabled)")
-	}
-
-	if vminfo.UEFI && espDiskIndex >= 0 && espDiskIndex != finalBootIndex {
-		migobj.logMessage(fmt.Sprintf("UEFI multi-disk: ESP on Disk %d, Root on Disk %d", espDiskIndex, finalBootIndex))
-	}
-
-	return finalBootIndex, finalOsPath, osRelease, espDiskIndex, nil
-}
-
-// validateLinuxOS checks if the detected Linux OS is supported
-func (migobj *Migrate) validateLinuxOS(osRelease string) error {
-	osDetected := strings.ToLower(strings.TrimSpace(osRelease))
-	utils.PrintLog(fmt.Sprintf("OS detected by guestfish: %s", osDetected))
-
-	supportedOS := []string{
-		"redhat", "red hat", "rhel", "centos", "scientific linux",
-		"oracle linux", "fedora", "sles", "sled", "opensuse",
-		"alt linux", "debian", "ubuntu", "rocky linux",
-		"suse linux enterprise server", "suse linux enterprise desktop", "alma linux",
-	}
-
-	for _, s := range supportedOS {
-		if strings.Contains(osDetected, s) {
-			utils.PrintLog("operating system compatibility check passed")
-			return nil
-		}
-	}
-
-	return errors.Errorf("unsupported OS detected by guestfish: %s", osDetected)
-}
-
-// handleWindowsBootDetection handles boot volume detection for Windows systems
-func (migobj *Migrate) handleWindowsBootDetection(vminfo vm.VMInfo, bootVolumeIndex int) (int, string, error) {
-	utils.PrintLog("operating system compatibility check passed")
-
-	var finalBootIndex int
-	var err error
-
-	if len(vminfo.VMDisks) == 0 {
-		return -1, "", errors.New("no disks present for VM; cannot detect boot disk")
-	}
-
-	utils.PrintLog("checking for bootable volume in case of LDM")
-	finalBootIndex, err = virtv2v.GetBootableVolumeIndex(vminfo.VMDisks)
-	if err != nil {
-		return -1, "", errors.Wrap(err, "Failed to get bootable volume index")
-	}
-
-	// Same guard as the Linux path: GetBootableVolumeIndex resolves via device-index,
-	// which can point at the libguestfs appliance disk (past the guest's disks).
-	// Fail the migration rather than indexing vminfo.VMDisks out of range.
-	if finalBootIndex < 0 || finalBootIndex >= len(vminfo.VMDisks) {
-		return -1, "", errors.Errorf(
-			"detected boot disk index %d is out of range for %d disk(s); aborting migration",
-			finalBootIndex, len(vminfo.VMDisks))
-	}
-
-	osRelease, err := virtv2v.GetWindowsVersion(vminfo.VMDisks, vminfo.VMDisks[finalBootIndex].Path)
-	if err != nil {
-		utils.PrintLog(fmt.Sprintf("Warning: Failed to detect Windows version: %v", err))
-		osRelease = "Windows (version unknown)"
-	}
-
-	utils.PrintLog(fmt.Sprintf("Windows OS detected: %s", osRelease))
-
-	return finalBootIndex, osRelease, nil
-}
-
-// performDiskConversion runs virt-v2v conversion on the boot disk
-func (migobj *Migrate) performDiskConversion(ctx context.Context, vminfo vm.VMInfo, bootVolumeIndex int, osPath, osRelease string, espDiskIndex int) error {
-
-	persisNetwork := utils.GetNetworkPersistance(ctx, migobj.K8sClient)
-	removeVMwareTools := utils.GetRemoveVMwareTools(ctx, migobj.K8sClient)
-
-	if !migobj.Convert {
-		return nil
-	}
-
-	firstbootscripts := []string{}
-	firstbootwinscripts := []virtv2v.FirstBootWindows{}
-	// Fix NTFS for Windows
-	if strings.ToLower(vminfo.OSType) == constants.OSFamilyWindows {
-		if err := virtv2v.NTFSFix(vminfo.VMDisks[bootVolumeIndex].Path); err != nil {
-			return errors.Wrap(err, "failed to run ntfsfix")
-		}
-		firstbootscripts = append(firstbootscripts, "Firstboot-Init-Windows")
-		firstbootwinscripts = append(firstbootwinscripts, virtv2v.FirstBootWindows{
-			Script: "Firstboot-Scheduler.ps1",
-			Async:  false,
-		})
-		if strings.ToLower(vminfo.OSType) == constants.OSFamilyWindows {
-			utils.PrintLog("Successfully added VirtIO PowerShell script to guest")
-			firstbootwinscripts = append(firstbootwinscripts, virtv2v.FirstBootWindows{
-				Script: "install-virtio-win12.ps1",
-				Async:  false,
-			})
-		}
-		if persisNetwork {
-			firstbootwinscripts = append(firstbootwinscripts, virtv2v.FirstBootWindows{
-				Script: "Orchestrate-NICRecovery.ps1",
-				Async:  true,
-			})
-		}
-		firstbootwinscripts = append(firstbootwinscripts, virtv2v.FirstBootWindows{
-			Script: "disk-online-fix.ps1",
-			Async:  true,
-		})
-		if removeVMwareTools {
-			firstbootwinscripts = append(firstbootwinscripts, virtv2v.FirstBootWindows{
-				Script: "vmware-tools-deletion.ps1",
-				Async:  true,
-			})
-		}
-		userFirstBootScripts, err := virtv2v.PushWindowsFirstBoot(vminfo.OSType)
-		if err != nil {
-			return err
-		}
-		for _, scriptName := range userFirstBootScripts {
-			firstbootwinscripts = append(firstbootwinscripts, virtv2v.FirstBootWindows{
-				Script: scriptName,
-				Async:  true,
-			})
-		}
-	}
-
-	// Add first boot scripts for RHEL family
-	if virtv2v.IsRHELFamily(osRelease) {
-		versionID := parseVersionID(osRelease)
-		if versionID == "" {
-			return errors.Errorf("failed to get version ID")
-		}
-		if !persisNetwork {
-
-			majorVersion, err := strconv.Atoi(strings.Split(versionID, ".")[0])
-			if err != nil {
-				return fmt.Errorf("failed to parse major version: %v", err)
-			}
-
-			if majorVersion >= 7 {
-				firstbootscriptname := "rhel_enable_dhcp"
-				firstbootscript := constants.RhelFirstBootScript
-				firstbootscripts = append(firstbootscripts, firstbootscriptname)
-
-				if err := virtv2v.AddFirstBootScript(firstbootscript, firstbootscriptname); err != nil {
-					return errors.Wrap(err, "failed to add first boot script")
-				}
-				utils.PrintLog("First boot script added successfully")
-			}
-		}
-	}
-
-	// Pre-conversion SUSE fixes: install the LVM mkinitrd wrapper so that
-	// virt-v2v's chroot call to mkinitrd can resolve /dev/<vg>/<lv> paths.
-	if virtv2v.IsSUSEFamily(osRelease) {
-		utils.PrintLog("SUSE guest detected: running pre-conversion FixLegacyMkinitrd")
-		if err := virtv2v.FixLegacyMkinitrd(vminfo.VMDisks); err != nil {
-			// Non-fatal: log and continue.  The conversion may still succeed
-			// on modern SUSE guests that use dracut, or if the root is not LVM.
-			utils.PrintLog(fmt.Sprintf("Warning: FixLegacyMkinitrd failed (continuing): %v", err))
-		} else {
-			utils.PrintLog("FixLegacyMkinitrd completed successfully")
-		}
-	}
-	// Inject VMware Tools cleanup script for Linux guests when requested
-	if removeVMwareTools && strings.ToLower(vminfo.OSType) == constants.OSFamilyLinux {
-		firstbootscriptname := "vmware_tools_cleanup"
-		firstbootscripts = append(firstbootscripts, firstbootscriptname)
-		scriptContent, err := os.ReadFile("/home/fedora/vmware-tools-cleanup.sh")
-		if err != nil {
-			return errors.Wrap(err, "failed to read VMware tools cleanup script")
-		}
-		if err := virtv2v.AddFirstBootScript(string(scriptContent), firstbootscriptname); err != nil {
-			return errors.Wrap(err, "failed to add VMware tools cleanup first boot script")
-		}
-		utils.PrintLog("VMware Tools cleanup script added for Linux firstboot")
-	}
-
-	// Run virt-v2v conversion
-	blockDriver := blockDriverFromMetadata(migobj.ImageMetadata)
-	utils.PrintLog(fmt.Sprintf("Starting virt-v2v conversion for VM %s (osPath=%s, osType=%s, blockDriver=%q)", vminfo.Name, osPath, vminfo.OSType, blockDriver))
-	if err := virtv2v.ConvertDisk(ctx, constants.XMLFileName, osPath, vminfo.OSType, migobj.Virtiowin, firstbootscripts, vminfo.VMDisks[bootVolumeIndex].Path, osRelease, blockDriver); err != nil {
-		return errors.Wrap(err, "failed to run virt-v2v")
-	}
-	utils.PrintLog("virt-v2v conversion completed successfully")
-
-	if strings.ToLower(vminfo.OSType) == constants.OSFamilyWindows {
-		if removeVMwareTools {
-			if err := virtv2v.RunOfflineVMwareCleanup(vminfo.VMDisks[bootVolumeIndex].Path); err != nil {
-				utils.PrintLog(fmt.Sprintf("WARNING: offline VMware Tools cleanup returned error: %v", err))
-			}
-		}
-
-		if err := virtv2v.InjectFirstBootScriptsFromStore(vminfo.VMDisks, vminfo.VMDisks[bootVolumeIndex].Path, firstbootwinscripts); err != nil {
-			return errors.Wrap(err, "failed to inject first boot scripts")
-		}
-	}
-
-	// Set volume as bootable
-	if err := migobj.Openstackclients.SetVolumeBootable(ctx, vminfo.VMDisks[bootVolumeIndex].OpenstackVol); err != nil {
-		return errors.Wrap(err, "failed to set volume as bootable")
-	}
-
-	// For UEFI multi-disk layouts, also mark the ESP disk as bootable
-	// This is required because OpenStack won't allow attaching a non-bootable volume with BootIndex=0
-	if vminfo.UEFI && espDiskIndex >= 0 && espDiskIndex != bootVolumeIndex {
-		migobj.logMessage(fmt.Sprintf("Marking ESP disk (Disk %d: %s) as bootable in OpenStack", espDiskIndex, vminfo.VMDisks[espDiskIndex].Name))
-		if err := migobj.Openstackclients.SetVolumeBootable(ctx, vminfo.VMDisks[espDiskIndex].OpenstackVol); err != nil {
-			return errors.Wrap(err, "failed to set ESP volume as bootable")
-		}
-	}
-
-	return nil
-}
-func (migobj *Migrate) configureWindowsNetwork(ctx context.Context, vminfo vm.VMInfo, bootVolumeIndex int, osRelease string) error {
-	persistNetwork := utils.GetNetworkPersistance(ctx, migobj.K8sClient)
-	if persistNetwork {
-		osType := strings.ToLower(vminfo.OSType)
-		if err := virtv2v.InjectMacToIps(vminfo.VMDisks, vminfo.VMDisks[bootVolumeIndex].Path, vminfo.GuestNetworks, vminfo.GatewayIP, vminfo.IPperMac, osType); err != nil {
-			return errors.Wrap(err, "failed to inject mac to ips")
-		}
-		utils.PrintLog("Mac to IP mapping injected successfully")
-		if err := virtv2v.InjectRestorationScript(vminfo.VMDisks, vminfo.VMDisks[bootVolumeIndex].Path); err != nil {
-			return errors.Wrap(err, "failed to inject restoration script")
-		}
-		utils.PrintLog("Restoration script injected successfully")
-	}
-	return nil
-}
-
-// configureLinuxNetwork handles network configuration for Linux systems
-func (migobj *Migrate) configureLinuxNetwork(ctx context.Context, vminfo vm.VMInfo, bootVolumeIndex int, osRelease string) error {
-	persisNetwork := utils.GetNetworkPersistance(ctx, migobj.K8sClient)
-	if persisNetwork {
-		osType := strings.ToLower(vminfo.OSType)
-		if err := virtv2v.InjectMacToIps(vminfo.VMDisks, vminfo.VMDisks[bootVolumeIndex].Path, vminfo.GuestNetworks, vminfo.GatewayIP, vminfo.IPperMac, osType); err != nil {
-			return errors.Wrap(err, "failed to inject mac to ips")
-		}
-		utils.PrintLog("Mac to ips injection completed successfully")
-		versionID := parseVersionID(osRelease)
-		if versionID == "" {
-			return errors.Errorf("failed to get version ID")
-		}
-		isNetplan := isNetplanSupported(versionID) && strings.Contains(osRelease, "ubuntu")
-		utils.PrintLog(fmt.Sprintf("Is netplan: %v", isNetplan))
-		utils.PrintLog("Running network persistence script")
-		if err := virtv2v.RunNetworkPersistence(vminfo.VMDisks, vminfo.VMDisks[bootVolumeIndex].Path, vminfo.OSType, isNetplan); err != nil {
-			utils.PrintLog(fmt.Sprintf("Warning: Network persistence script failed: %v", err))
-		} else {
-			utils.PrintLog("Network persistence script executed successfully")
-		}
-	} else {
-		if strings.Contains(osRelease, "ubuntu") {
-			return migobj.configureUbuntuNetwork(vminfo, bootVolumeIndex, osRelease)
-		}
-
-		if virtv2v.IsRHELFamily(osRelease) {
-			return migobj.configureRHELNetwork(vminfo, bootVolumeIndex, osRelease)
-		}
-	}
-
-	return nil
-}
-
-// configureUbuntuNetwork handles Ubuntu-specific network configuration
-func (migobj *Migrate) configureUbuntuNetwork(vminfo vm.VMInfo, bootVolumeIndex int, osRelease string) error {
-	versionID := parseVersionID(osRelease)
-	utils.PrintLog(fmt.Sprintf("Version ID: %s", versionID))
-
-	if versionID == "" {
-		return errors.Errorf("failed to get version ID")
-	}
-
-	if isNetplanSupported(versionID) {
-		utils.PrintLog("Adding wildcard netplan")
-		if migobj.isSimpleNetwork {
-			utils.PrintLog("L2 network detected adding l2 wildcard")
-			if err := virtv2v.AddWildcardNetplanForL2(vminfo.VMDisks, vminfo.VMDisks[bootVolumeIndex].Path); err != nil {
-				return errors.Wrap(err, "failed to add l2 wildcard netplan")
-			}
-		} else {
-			if err := virtv2v.AddWildcardNetplan(vminfo.VMDisks, vminfo.VMDisks[bootVolumeIndex].Path, vminfo.GuestNetworks, vminfo.GatewayIP, vminfo.IPperMac); err != nil {
-				return errors.Wrap(err, "failed to add wildcard netplan")
-			}
-		}
-		utils.PrintLog("Wildcard netplan added successfully")
-		return nil
-	}
-
-	return migobj.addUdevRulesForUbuntu(vminfo, bootVolumeIndex)
-}
-
-// addUdevRulesForUbuntu adds udev rules for older Ubuntu versions
-func (migobj *Migrate) addUdevRulesForUbuntu(vminfo vm.VMInfo, bootVolumeIndex int) error {
-	utils.PrintLog("Ubuntu version does not support netplan, going to use udev rules")
-
-	interfaces, err := virtv2v.GetNetworkInterfaceNames(vminfo.VMDisks[bootVolumeIndex].Path)
-	if err != nil {
-		return errors.Wrap(err, "failed to get network interface names")
-	}
-
-	if len(interfaces) == 0 {
-		log.Printf("Failed to get network interface names, cannot add udev rules, network might not come up post migration, please check the network configuration post migration")
-		return nil
-	}
-
-	utils.PrintLog("Adding udev rules")
-	utils.PrintLog(fmt.Sprintf("Interfaces: %v", interfaces))
-
-	macs := []string{}
-	for _, nic := range vminfo.NetworkInterfaces {
-		macs = append(macs, nic.MAC)
-	}
-	utils.PrintLog(fmt.Sprintf("MACs: %v", macs))
-
-	if err := virtv2v.AddUdevRules(vminfo.VMDisks, vminfo.VMDisks[bootVolumeIndex].Path, interfaces, macs); err != nil {
-		log.Printf(`Warning Failed to add udev rules: %s, incase of interface name mismatch,
-                        network might not come up post migration, please check the network configuration post migration`, err)
-		log.Println("Continuing with migration")
-	}
-
-	return nil
-}
-
-func (migobj *Migrate) configureRHELNetwork(vminfo vm.VMInfo, bootVolumeIndex int, osRelease string) error {
-	versionID := parseVersionID(osRelease)
-	majorVersion, err := strconv.Atoi(strings.Split(versionID, ".")[0])
-	if err != nil {
-		return fmt.Errorf("failed to parse major version: %v", err)
-	}
-
-	if majorVersion < 7 {
-		diskPath := vminfo.VMDisks[bootVolumeIndex].Path
-		if err := DetectAndHandleNetwork(diskPath, osRelease, vminfo); err != nil {
-			utils.PrintLog(fmt.Sprintf(`Warning: Failed to handle network: %v,Continuing with migration,
-                    network might not come up post migration, please check the network configuration post migration`, err))
-		}
-	}
-
-	return nil
-}
-
-// blockDriverFromMetadata returns the virt-v2v --block-driver value that
-// matches the hw_disk_bus / hw_scsi_model image properties the caller intends
-// to set on the migrated volume.  The two values must agree: if User wants to use
-// virtio-scsi controller (hw_disk_bus=scsi, hw_scsi_model=virtio-scsi)
-// virt-v2v must make vioscsi boot-critical; or else virtio-blk
-// (hw_disk_bus=virtio, the default) virt-v2v must make viostor boot-critical.
-// Returning "" lets virt-v2v use its built-in default (virtio-blk / viostor).
-func blockDriverFromMetadata(metadata map[string]string) string {
-	if metadata["hw_disk_bus"] == "scsi" && metadata["hw_scsi_model"] == "virtio-scsi" {
-		return "virtio-scsi"
-	}
-	return ""
-}
-
-func (migobj *Migrate) ConvertVolumes(ctx context.Context, vminfo vm.VMInfo) (int, error) {
-	migobj.logMessage("Converting disk")
-
-	// Step 1: Determine boot command based on OS type
-	getBootCommand := migobj.getBootCommand(vminfo.OSType)
-
-	// Step 2: Attach all volumes
-	if err := migobj.attachAllVolumes(ctx, &vminfo); err != nil {
-		return -1, err
-	}
-
-	// Step 3: Generate XML configuration for conversion
-	if err := vmutils.GenerateXMLConfig(vminfo); err != nil {
-		return -1, errors.Wrap(err, "failed to generate XML")
-	}
-
-	// Step 3.5: Get vjailbreak settings
-	vjailbreakSettings, err := k8sutils.GetVjailbreakSettings(ctx, migobj.K8sClient)
-	if err != nil {
-		return -1, errors.Wrap(err, "failed to get vjailbreak settings")
-	}
-
-	// Step 4: Detect boot volume
-	bootVolumeIndex, osPath, err := migobj.detectBootVolume(vminfo, getBootCommand)
-	if err != nil {
-		return -1, err
-	}
-
-	// Step 5: Handle OS-specific detection and validation
-	var osRelease string
-	var espDiskIndex int = -1
-	osType := strings.ToLower(vminfo.OSType)
-
-	switch osType {
-	case constants.OSFamilyLinux:
-		bootVolumeIndex, osPath, osRelease, espDiskIndex, err = migobj.handleLinuxOSDetection(vminfo, bootVolumeIndex, osPath, vjailbreakSettings.AutoFstabUpdate)
-		if err != nil {
-			return -1, err
-		}
-
-	case constants.OSFamilyWindows:
-		bootVolumeIndex, osRelease, err = migobj.handleWindowsBootDetection(vminfo, bootVolumeIndex)
-		if err != nil {
-			return -1, err
-		}
-
-	default:
-		return -1, errors.Errorf("unsupported OS type: %s", vminfo.OSType)
-	}
-
-	// Step 6: Validate boot volume was found
-	if bootVolumeIndex == -1 {
-		return -1, errors.Errorf("boot volume not found, cannot create target VM")
-	}
-
-	// Step 7: Mark boot volume
-	utils.PrintLog(fmt.Sprintf("Boot disk selected: Disk %d (%s)", bootVolumeIndex, vminfo.VMDisks[bootVolumeIndex].Name))
-	vminfo.VMDisks[bootVolumeIndex].Boot = true
-
-	// Step 8: Apply merged VolumeImageProfile metadata to the boot volume. Nova/libvirt
-	// only read volume_image_metadata from the root disk, so we scope this to the boot volume.
-	if len(migobj.ImageMetadata) > 0 {
-		bootVol := vminfo.VMDisks[bootVolumeIndex].OpenstackVol
-		if bootVol != nil {
-			if err := migobj.Openstackclients.ApplyBootVolumeImageMetadata(ctx, bootVol, migobj.ImageMetadata); err != nil {
-				return -1, errors.Wrap(err, "failed to apply VolumeImageProfile metadata to boot volume")
-			}
-			migobj.logMessage(fmt.Sprintf("Applied %d image metadata key(s) from VolumeImageProfiles to boot volume %s: %v",
-				len(migobj.ImageMetadata), bootVol.ID, migobj.ImageMetadata))
-		}
-	}
-
-	// Step 9: Perform disk conversion
-	if err := migobj.performDiskConversion(ctx, vminfo, bootVolumeIndex, osPath, osRelease, espDiskIndex); err != nil {
-		return -1, err
-	}
-
-	// Step 10: Configure network for Linux systems
-	if osType == constants.OSFamilyLinux {
-		if err := migobj.configureLinuxNetwork(ctx, vminfo, bootVolumeIndex, osRelease); err != nil {
-			return -1, err
-		}
-	} else if osType == constants.OSFamilyWindows {
-		if err := migobj.configureWindowsNetwork(ctx, vminfo, bootVolumeIndex, osRelease); err != nil {
-			return -1, err
-		}
-	}
-
-	// Step 11: Detach all volumes
-	if err := migobj.DetachAllVolumes(ctx, vminfo); err != nil {
-		return -1, errors.Wrap(err, "Failed to detach all volumes from VM")
-	}
-
-	migobj.logMessage("Successfully converted disk")
-	return espDiskIndex, nil
-}
-
-// DetectAndHandleNetwork: Checks if RHEL family, then detects NM presence offline.
-// If NM (nmcli exists), injects first-boot nmcli script for DHCP force.
-// If not, adds udev rules to pin names without forcing DHCP.
-func DetectAndHandleNetwork(diskPath string, osRelease string, vmInfo vm.VMInfo) error {
-
-	// No NM: Add udev rules to pin names
-	interfaces, err := virtv2v.GetInterfaceNames(diskPath)
-	if err != nil {
-		utils.PrintLog(fmt.Sprintf("Warning: Failed to get interfaces: %v", err))
-	}
-	if len(interfaces) == 0 {
-		utils.PrintLog(`No network interfaces found, cannot add udev rules, network might not
-            come up post migration, please check the network configuration post migration`)
-		return nil
-	}
-	macs := []string{}
-	// By default the network interfaces macs are in the same order as the interfaces
-	for _, nic := range vmInfo.NetworkInterfaces {
-		macs = append(macs, nic.MAC)
-	}
-	utils.PrintLog(fmt.Sprintf("Interfaces: %v", interfaces))
-	utils.PrintLog(fmt.Sprintf("MACs: %v", macs))
-	if len(interfaces) != len(macs) {
-		utils.PrintLog("Mismatch between number of interfaces and MACs")
-		return fmt.Errorf("mismatch between number of interfaces and MACs")
-	}
-	// Add udev rules to pin names without forcing DHCP
-	utils.PrintLog("Adding udev rules to pin interface names")
-
-	// This will ensure that the network interfaces are named consistently after migration
-	// and they get the correct IP address.
-	// This is important because RHEL family uses NetworkManager by default and it does not
-	// automatically configure the network interfaces to use DHCP after migration.
-	// So we need to add udev rules to pin the names of the network interfaces
-	// to the MAC addresses so that they are consistent after migration.
-	// This will ensure that the network interfaces are named consistently after migration
-	// and they get the correct IP address.
-	err = virtv2v.AddUdevRules([]vm.VMDisk{{Path: diskPath}}, diskPath, interfaces, macs)
-	if err != nil {
-		utils.PrintLog(fmt.Sprintf("Warning: Failed to add udev: %v", err))
-	}
-	return nil
-}
-
-func (migobj *Migrate) CreateTargetInstance(ctx context.Context, vminfo vm.VMInfo, networkids, portids []string, ipaddresses []string, espDiskIndex int) error {
-	migobj.logMessage("Creating target instance")
-	openstackops := migobj.Openstackclients
-	var flavor *flavors.Flavor
-	var err error
-
-	if migobj.TargetFlavorId != "" {
-		flavor, err = openstackops.GetFlavor(ctx, migobj.TargetFlavorId)
-		if err != nil {
-			return errors.Wrap(err, "failed to get OpenStack flavor")
-		}
-		if utils.IsHotplugFlavor(flavor) {
-			migobj.logMessage(fmt.Sprintf("Assigned flavor %s is a hotplug base flavor (0 vCPU, 0 RAM); VM will be created with hotplug metadata", flavor.Name))
-		}
-	} else {
-		flavor, err = openstackops.GetClosestFlavour(ctx, vminfo.CPU, vminfo.Memory)
-		if err != nil {
-			return errors.Wrap(err, "failed to get closest OpenStack flavor")
-		}
-		utils.PrintLog(fmt.Sprintf("Closest OpenStack flavor: %s: CPU: %dvCPUs\tMemory: %dMB\n", flavor.Name, flavor.VCPUs, flavor.RAM))
-	}
-	// Get security group IDs
-	securityGroupIDs, err := openstackops.GetSecurityGroupIDs(ctx, migobj.SecurityGroups, migobj.TenantName)
-	if err != nil {
-		return errors.Wrap(err, "failed to resolve security group names to IDs")
-	}
-	utils.PrintLog(fmt.Sprintf("Using security group IDs: %v", securityGroupIDs))
-
-	if migobj.ServerGroup != "" {
-		utils.PrintLog(fmt.Sprintf("Using server group ID: %s", migobj.ServerGroup))
-	} else {
-		utils.PrintLog("No server group specified - VMs will be placed based on default scheduling")
-	}
-
-	// Get vjailbreak settings
-	vjailbreakSettings, err := k8sutils.GetVjailbreakSettings(context.Background(), migobj.K8sClient)
-	if err != nil {
-		return errors.Wrap(err, "failed to get vjailbreak settings")
-	}
-	utils.PrintLog(fmt.Sprintf("Fetched vjailbreak settings for VM active wait retry limit: %d, VM active wait interval seconds: %d", vjailbreakSettings.VMActiveWaitRetryLimit, vjailbreakSettings.VMActiveWaitIntervalSeconds))
-
-	// Create a new VM in OpenStack
-	if len(migobj.TargetMetadata) > 0 {
-		vminfo.TargetMetadata = migobj.TargetMetadata
-		migobj.logMessage(fmt.Sprintf("Applying %d instance metadata entries (preserved source tags/custom metadata) to target VM", len(migobj.TargetMetadata)))
-	}
-	newVM, err := openstackops.CreateVM(ctx, flavor, networkids, portids, vminfo, migobj.TargetAvailabilityZone, securityGroupIDs, migobj.ServerGroup, *vjailbreakSettings, espDiskIndex)
-	if err != nil {
-		return errors.Wrap(err, "failed to create VM")
-	}
-
-	// Wait for VM to become active
-	for i := 0; i < vjailbreakSettings.VMActiveWaitRetryLimit; i++ {
-		migobj.logMessage(fmt.Sprintf("Waiting for VM to become active: %d/%d retries\n", i+1, vjailbreakSettings.VMActiveWaitRetryLimit))
-		active, err := openstackops.WaitUntilVMActive(ctx, newVM.ID)
-		if err != nil {
-			return errors.Wrap(err, "failed to wait for VM to become active")
-		}
-		if active {
-			break
-		}
-		if i == vjailbreakSettings.VMActiveWaitRetryLimit-1 {
-			return errors.Errorf("VM is not active after %d retries", vjailbreakSettings.VMActiveWaitRetryLimit)
-		}
-		time.Sleep(time.Duration(vjailbreakSettings.VMActiveWaitIntervalSeconds) * time.Second)
-	}
-
-	migobj.logMessage(fmt.Sprintf("VM created successfully: ID: %s", newVM.ID))
-
-	if migobj.PerformHealthChecks {
-		err = migobj.HealthCheck(vminfo, ipaddresses)
-		if err != nil {
-			migobj.logMessage(fmt.Sprintf("Health Check failed: %s", err))
-		}
-	} else {
-		migobj.logMessage("Skipping Health Checks")
-	}
-
-	return nil
-}
-
-// parseVersionID parses the VERSION_ID from /etc/os-release or /etc/redhat-release format.
-// It returns the version ID as a string, or an empty string if not found.
-func parseVersionID(osRelease string) string {
-	osRelease = strings.TrimSpace(osRelease)
-
-	// Key-value style (os-release, SuSE-release, etc.)
-	if strings.Contains(osRelease, "=") {
-		var version, patchlevel string
-		for _, line := range strings.Split(osRelease, "\n") {
-			kv := strings.SplitN(line, "=", 2)
-			if len(kv) != 2 {
-				continue
-			}
-			key := strings.TrimSpace(strings.ToUpper(kv[0]))
-			val := strings.TrimSpace(strings.Trim(kv[1], `"`)) // Remove quotes and spaces
-			switch key {
-			case "VERSION_ID":
-				return val
-			case "VERSION":
-				version = val
-			case "PATCHLEVEL":
-				patchlevel = val
-			}
-		}
-		// If it's SLES style, combine VERSION + PATCHLEVEL if available
-		if version != "" {
-			if patchlevel != "" {
-				return version + "." + patchlevel
-			}
-			return version
-		}
-	} else {
-		// /etc/redhat-release style
-		re := regexp.MustCompile(`release\s+([0-9]+(\.[0-9]+)?)`)
-		matches := re.FindStringSubmatch(strings.ToLower(osRelease))
-		if len(matches) > 1 {
-			return matches[1]
-		}
-	}
-
-	return ""
-}
-
-func isNetplanSupported(version string) bool {
-	parts := strings.Split(version, ".")
-	if len(parts) < 2 {
-		log.Printf("Warning: unexpected VERSION_ID format: %q", version)
-		return true // assume modern if uncertain
-	}
-
-	major, err1 := strconv.Atoi(parts[0])
-	minor, err2 := strconv.Atoi(parts[1])
-	if err1 != nil || err2 != nil {
-		log.Printf("Warning: failed to parse VERSION_ID %q: %v %v", version, err1, err2)
-		return true
-	}
-
-	// Compare with 17.10
-	if major > 17 {
-		return true
-	}
-	if major == 17 && minor >= 10 {
-		return true
-	}
-	return false
-}
-
-func (migobj *Migrate) pingVM(ips []string) error {
-	for _, ip := range ips {
-		migobj.logMessage(fmt.Sprintf("Pinging VM: %s", ip))
-		pinger, err := probing.NewPinger(ip)
-		if err != nil {
-			return errors.Wrap(err, "failed to create pinger")
-		}
-		pinger.Count = 1
-		pinger.Timeout = time.Second * 10
-		err = pinger.Run()
-		if err != nil {
-			return errors.Wrap(err, "failed to run pinger")
-		}
-		if pinger.Statistics().PacketLoss == 0 {
-			migobj.logMessage("Ping succeeded")
-		} else {
-			return errors.Errorf("Ping failed")
-		}
-	}
-	return nil
-}
-
-func (migobj *Migrate) checkHTTPGet(ips []string, port string) error {
-	var client *http.Client
-	vjbNet := netutils.NewVjbNet()
-	if migobj.Insecure {
-		vjbNet.Insecure = true
-	}
-	if vjbNet.CreateSecureHTTPClient() == nil {
-		client = vjbNet.GetClient()
-	} else {
-		return errors.Errorf("Both HTTP and HTTPS failed ")
-	}
-	for _, ip := range ips {
-		// Try HTTP first
-		httpURL := fmt.Sprintf("http://%s:%s", ip, port)
-		if err := migobj.tryConnection(client, httpURL); err == nil {
-			migobj.logMessage("HTTP succeeded")
-			continue // Success with HTTP, move to next IP
-		}
-
-		// If HTTP fails, try HTTPS
-		httpsURL := fmt.Sprintf("https://%s:%s", ip, port)
-		if err := migobj.tryConnection(client, httpsURL); err == nil {
-			migobj.logMessage("HTTPS succeeded")
-			continue // Success with HTTPS, move to next IP
-		}
-
-		// Both HTTP and HTTPS failed
-		return errors.Errorf("Both HTTP and HTTPS failed for %s:%s", ip, port)
-	}
-
-	return nil
-}
-
-func (migobj *Migrate) tryConnection(client *http.Client, url string) error {
-	resp, err := client.Get(url)
-	if err != nil {
-		migobj.logMessage(fmt.Sprintf("GET failed for %s: %v", url, err))
-		return errors.Wrap(err, "failed to get url")
-	}
-	defer resp.Body.Close()
-
-	migobj.logMessage(fmt.Sprintf("GET response for %s: %d", url, resp.StatusCode))
-
-	if resp.StatusCode != http.StatusOK {
-		return errors.Errorf("GET returned non-OK status for %s: %d", url, resp.StatusCode)
-	}
-
-	return nil
-}
-
-func (migobj *Migrate) HealthCheck(vminfo vm.VMInfo, ips []string) error {
-	migobj.logMessage("Performing Health Checks")
-	healthChecks := make(map[string]bool)
-	healthChecks["Ping"] = false
-	healthChecks["HTTP Get"] = false
-	for i := 0; i < 10; i++ {
-		migobj.logMessage(fmt.Sprintf("Health Check Attempt %d", i+1))
-		// 1. Ping
-		if !healthChecks["Ping"] {
-			err := migobj.pingVM(ips)
-			if err != nil {
-				migobj.logMessage(fmt.Sprintf("Ping(s) failed: %s", err))
-			} else {
-				healthChecks["Ping"] = true
-			}
-		}
-		// 2. HTTP GET check
-		if !healthChecks["HTTP Get"] {
-			err := migobj.checkHTTPGet(ips, migobj.HealthCheckPort)
-			if err != nil {
-				migobj.logMessage(fmt.Sprintf("HTTP Get failed: %s", err))
-			} else {
-				healthChecks["HTTP Get"] = true
-			}
-		}
-		if healthChecks["Ping"] && healthChecks["HTTP Get"] {
-			break
-		}
-		migobj.logMessage("Waiting for 60 seconds before retrying health checks")
-		time.Sleep(60 * time.Second)
-	}
-	for key, value := range healthChecks {
-		if !value {
-			migobj.logMessage(fmt.Sprintf("Health Check %s failed", key))
-		} else {
-			migobj.logMessage(fmt.Sprintf("Health Check %s succeeded", key))
-		}
-	}
-	return nil
 }
 
 func (migobj *Migrate) gracefulTerminate(ctx context.Context, vminfo vm.VMInfo, cancel context.CancelFunc) {
@@ -2293,431 +394,6 @@ func (migobj *Migrate) cleanup(ctx context.Context, vminfo vm.VMInfo, message st
 	} else if vcenterSettings != nil && !vcenterSettings.CleanupPortsAfterMigrationFailure {
 		migobj.logMessage("Cleanup ports after migration failure is disabled, ports will not be deleted")
 	}
-
-	return nil
-}
-
-func (migobj *Migrate) DeleteAllPorts(ctx context.Context, portids []string) error {
-	migobj.logMessage("Deleting all ports")
-	openstackops := migobj.Openstackclients
-	var deletionErrors []error
-	successCount := 0
-
-	for _, portID := range portids {
-		err := openstackops.DeletePort(ctx, portID)
-		if err != nil {
-			utils.PrintLog(fmt.Sprintf("Failed to delete port %s: %s\n", portID, err))
-			deletionErrors = append(deletionErrors, errors.Wrapf(err, "failed to delete port %s", portID))
-		} else {
-			utils.PrintLog(fmt.Sprintf("Successfully deleted port %s\n", portID))
-			successCount++
-		}
-	}
-
-	if len(deletionErrors) > 0 {
-		migobj.logMessage(fmt.Sprintf("Port deletion completed with errors: %d succeeded, %d failed out of %d total", successCount, len(deletionErrors), len(portids)))
-		// Combine all errors into a single error message
-		errMsg := fmt.Sprintf("failed to delete %d port(s):", len(deletionErrors))
-		for _, err := range deletionErrors {
-			errMsg += fmt.Sprintf("\n  - %s", err.Error())
-		}
-		return errors.New(errMsg)
-	}
-
-	migobj.logMessage(fmt.Sprintf("Successfully deleted all %d ports", successCount))
-	return nil
-}
-
-// ReservePortsForVM reserves ports for every VM NIC: reuseExistingPorts if the
-// user pre-created ports, otherwise createPortsForNetworks makes new ones.
-func (migobj *Migrate) ReservePortsForVM(ctx context.Context, vminfo *vm.VMInfo) ([]string, []string, []string, error) {
-	migobj.isSimpleNetwork = false
-
-	securityGroupIDs, err := migobj.Openstackclients.GetSecurityGroupIDs(ctx, migobj.SecurityGroups, migobj.TenantName)
-	if err != nil {
-		return nil, nil, nil, errors.Wrap(err, "failed to resolve security group names to IDs")
-	}
-	utils.PrintLog(fmt.Sprintf("Using provided security group IDs %v", securityGroupIDs))
-
-	if migobj.ServerGroup != "" {
-		utils.PrintLog(fmt.Sprintf("Server group ID for VM placement: %s", migobj.ServerGroup))
-	}
-
-	if len(migobj.Networkports) != 0 {
-		return migobj.reuseExistingPorts(ctx)
-	}
-	return migobj.createPortsForNetworks(ctx, vminfo, securityGroupIDs)
-}
-
-// reuseExistingPorts handles the pre-created-ports flow: migobj.Networkports
-// holds one OpenStack port ID per NIC, created outside this migration.
-func (migobj *Migrate) reuseExistingPorts(ctx context.Context) ([]string, []string, []string, error) {
-	if len(migobj.Networkports) != len(migobj.Networknames) {
-		return nil, nil, nil, errors.Errorf("number of network ports does not match number of network names")
-	}
-
-	networkids := []string{}
-	portids := []string{}
-	ipaddresses := []string{}
-	for _, portID := range migobj.Networkports {
-		retrPort, err := migobj.Openstackclients.GetPort(ctx, portID)
-		if err != nil {
-			return nil, nil, nil, errors.Wrap(err, "failed to get port")
-		}
-		networkids = append(networkids, retrPort.NetworkID)
-		portids = append(portids, retrPort.ID)
-		for _, fixedIP := range retrPort.FixedIPs {
-			ipaddresses = append(ipaddresses, fixedIP.IPAddress)
-		}
-	}
-	return networkids, portids, ipaddresses, nil
-}
-
-// createPortsForNetworks handles the create-new-ports flow: one port per
-// entry in migobj.Networknames, in NIC order.
-func (migobj *Migrate) createPortsForNetworks(ctx context.Context, vminfo *vm.VMInfo, securityGroupIDs []string) ([]string, []string, []string, error) {
-	networkids := []string{}
-	portids := []string{}
-	ipaddresses := []string{}
-
-	// subnetPortIndex is shared across every NIC of this VM: it's how
-	// distinct port names get assigned when two or more NICs land on the same
-	// subnet (see GetCreateOpts/buildPortName), so it must persist across the
-	// whole loop, not be recreated per NIC.
-	subnetPortIndex := make(map[string]int)
-
-	for idx, networkname := range migobj.Networknames {
-		network, err := migobj.Openstackclients.GetNetwork(ctx, networkname)
-		if err != nil {
-			return nil, nil, nil, errors.Wrap(err, "failed to get network")
-		}
-		if network == nil {
-			return nil, nil, nil, errors.Errorf("network not found")
-		}
-
-		isSimpleNetwork, err := migobj.Openstackclients.GetIsSimpleNetwork(ctx, network.ID)
-		if err != nil {
-			return nil, nil, nil, errors.Wrap(err, "failed to check if network is L2")
-		}
-		migobj.isSimpleNetwork = migobj.isSimpleNetwork || isSimpleNetwork
-
-		override, err := resolveNICOverride(migobj.NetworkOverrides, idx)
-		if err != nil {
-			return nil, nil, nil, err
-		}
-
-		mac := vminfo.Mac[idx]
-		detectedIPs := logAndCollectDetectedIPs(vminfo, mac)
-		loggedIPs := applyPreserveIPOverride(vminfo, idx, mac, override, detectedIPs, migobj.FallbackToDHCP)
-		mac = applyPreserveMACOverride(vminfo, idx, mac, override.preserveMAC)
-		utils.PrintLog(fmt.Sprintf("Using IPs for MAC %s: %v", vminfo.Mac[idx], loggedIPs))
-
-		port, err := migobj.createPort(ctx, network, mac, vminfo, securityGroupIDs, subnetPortIndex)
-		if err != nil {
-			return nil, nil, nil, err
-		}
-		syncIPperMacFromPort(vminfo, mac, port)
-
-		networkids = append(networkids, network.ID)
-		portids = append(portids, port.ID)
-		for _, fixedIP := range port.FixedIPs {
-			ipaddresses = append(ipaddresses, fixedIP.IPAddress)
-		}
-	}
-	utils.PrintLog(fmt.Sprintf("Gateways : %v", vminfo.GatewayIP))
-	return networkids, portids, ipaddresses, nil
-}
-
-// createPort creates a single OpenStack port on network for the given
-// (possibly overridden) MAC, using vminfo.IPperMac to determine fixed IPs.
-// subnetPortIndex must be the same map across every NIC of this VM (see
-// createPortsForNetworks) so NICs sharing a subnet get distinct port names.
-func (migobj *Migrate) createPort(ctx context.Context, network *networks.Network, mac string, vminfo *vm.VMInfo, securityGroupIDs []string, subnetPortIndex map[string]int) (*ports.Port, error) {
-	port, err := migobj.Openstackclients.ValidateAndCreatePort(ctx, network, mac, vminfo.IPperMac, vminfo.Name, securityGroupIDs, migobj.FallbackToDHCP, vminfo.GatewayIP, subnetPortIndex)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to create port group")
-	}
-	addressesOfPort := []string{}
-	for _, fixedIP := range port.FixedIPs {
-		addressesOfPort = append(addressesOfPort, fixedIP.IPAddress)
-	}
-	utils.PrintLog(fmt.Sprintf("Port created successfully: MAC:%s IP:%s and Security Groups:%v\n", port.MACAddress, addressesOfPort, securityGroupIDs))
-	return port, nil
-}
-
-// nicOverride is the resolved per-NIC configuration for preserveIP/preserveMAC
-// and any user-assigned replacement IP, defaulting to "preserve everything".
-type nicOverride struct {
-	preserveIP     bool
-	preserveMAC    bool
-	userAssignedIP []string
-}
-
-// resolveNICOverride resolves the NICOverride (if any) for interface idx
-// against the "preserve everything" default; a nil field means "not set".
-func resolveNICOverride(overrides []NICOverride, idx int) (nicOverride, error) {
-	result := nicOverride{preserveIP: true, preserveMAC: true}
-
-	for _, override := range overrides {
-		if override.InterfaceIndex != idx {
-			continue
-		}
-		if override.PreserveIP != nil {
-			result.preserveIP = *override.PreserveIP
-		}
-		if override.PreserveMAC != nil {
-			result.preserveMAC = *override.PreserveMAC
-		}
-		if override.UserAssignedIP != "" {
-			for _, ip := range strings.Split(override.UserAssignedIP, ",") {
-				if trimmedIP := strings.TrimSpace(ip); trimmedIP != "" {
-					result.userAssignedIP = append(result.userAssignedIP, trimmedIP)
-				}
-			}
-		}
-		break
-	}
-
-	if len(result.userAssignedIP) > 1 {
-		return nicOverride{}, errors.Errorf("multiple user assigned IPs not supported for an interface")
-	}
-	return result, nil
-}
-
-// logAndCollectDetectedIPs logs and returns the IPs VMware Tools reported for
-// mac, or nil if none were detected.
-func logAndCollectDetectedIPs(vminfo *vm.VMInfo, mac string) []string {
-	detectedIPs := vminfo.IPperMac[mac]
-	if len(detectedIPs) == 0 {
-		return nil
-	}
-	ippm := make([]string, 0, len(detectedIPs))
-	for _, detectedIP := range detectedIPs {
-		ippm = append(ippm, detectedIP.IP)
-	}
-	utils.PrintLog(fmt.Sprintf("Detected IPs from VMware Tools for MAC %s: %v", mac, detectedIPs))
-	return ippm
-}
-
-// applyPreserveIPOverride replaces vminfo.IPperMac[mac] with the user-assigned
-// IP, or clears it, when preserveIP is false; returns the IPs for logging.
-// If no custom IP is given, fallbackToDHCP picks nil (auto-allocate) vs an
-// empty slice (no fixed IPs) for GetCreateOpts.
-func applyPreserveIPOverride(vminfo *vm.VMInfo, idx int, mac string, override nicOverride, detectedIPs []string, fallbackToDHCP bool) []string {
-	if override.preserveIP {
-		return detectedIPs
-	}
-
-	if len(override.userAssignedIP) > 0 {
-		entries := make([]vm.IpEntry, 0, len(override.userAssignedIP))
-		for _, ip := range override.userAssignedIP {
-			entries = append(entries, vm.IpEntry{IP: ip, Prefix: 0})
-		}
-		vminfo.IPperMac[mac] = entries
-		utils.PrintLog(fmt.Sprintf("NIC[%d]: preserveIP=false, using user-assigned custom IP for MAC %s", idx, mac))
-		return override.userAssignedIP
-	}
-
-	if fallbackToDHCP {
-		utils.PrintLog(fmt.Sprintf("NIC[%d]: preserveIP=false, no custom IP for MAC %s, fallbackToDHCP=true — port will auto-allocate an IP", idx, mac))
-		vminfo.IPperMac[mac] = nil
-		return detectedIPs
-	}
-
-	utils.PrintLog(fmt.Sprintf("NIC[%d]: preserveIP=false, no custom IP for MAC %s, fallbackToDHCP=false — port will have no fixed IPs", idx, mac))
-	vminfo.IPperMac[mac] = []vm.IpEntry{}
-	return detectedIPs
-}
-
-// applyPreserveMACOverride moves mac's IPs to the "" key and returns "" so
-// OpenStack generates a new MAC, when preserveMAC is false. The real MAC
-// OpenStack assigns isn't known yet at this point in the flow — see
-// syncIPperMacFromPort, which re-keys "" to the real MAC once the port is
-// created.
-func applyPreserveMACOverride(vminfo *vm.VMInfo, idx int, mac string, preserveMAC bool) string {
-	if preserveMAC {
-		return mac
-	}
-	utils.PrintLog(fmt.Sprintf("NIC[%d]: preserveMAC=false for MAC %s — OpenStack will generate a new MAC", idx, mac))
-	vminfo.IPperMac[""] = vminfo.IPperMac[mac]
-	delete(vminfo.IPperMac, mac)
-	return ""
-}
-
-// syncIPperMacFromPort reconciles vminfo.IPperMac with the port OpenStack
-// actually created, keyed by placeholderMAC (the MAC passed to createPort:
-// the original MAC if preserveMAC=true, or "" if preserveMAC=false). Two
-// situations need reconciling once the real port exists:
-//
-//   - preserveMAC=false: OpenStack assigns a new MAC, but the entry is still
-//     sitting under the "" placeholder key. It's moved as-is to the real
-//     MAC (port.MACAddress) so MAC-keyed guest-config code (AddWildcardNetplan,
-//     the network-persistence script) can match it against the real NIC.
-//   - preserveIP=false + fallbackToDHCP=true: applyPreserveIPOverride left
-//     the entry nil so OpenStack would auto-allocate an IP instead of
-//     getting an explicit empty port. That nil is filled in with whatever
-//     IP(s) the port actually ended up with (possibly none, e.g. on an
-//     L2 network with no subnets).
-//
-// A NIC whose entries were already populated (preserveIP=true, or a custom
-// IP) is left untouched by the fill step: the port was created with exactly
-// that IP, so rebuilding from port.FixedIPs would risk losing a preserved
-// subnet prefix for no benefit.
-func syncIPperMacFromPort(vminfo *vm.VMInfo, placeholderMAC string, port *ports.Port) {
-	realMAC := port.MACAddress
-	if realMAC == "" {
-		// Defensive: don't lose the entry if a backend ever fails to echo a MAC.
-		realMAC = placeholderMAC
-	}
-
-	if realMAC != placeholderMAC {
-		vminfo.IPperMac[realMAC] = vminfo.IPperMac[placeholderMAC]
-		delete(vminfo.IPperMac, placeholderMAC)
-	}
-
-	if vminfo.IPperMac[realMAC] == nil {
-		// These IPs came from a live OpenStack auto-allocation, not a
-		// preserved/custom static IP — mark them so guest-config writers
-		// configure a real DHCP client instead of pinning the IP statically.
-		entries := make([]vm.IpEntry, 0, len(port.FixedIPs))
-		for _, fixedIP := range port.FixedIPs {
-			entries = append(entries, vm.IpEntry{IP: fixedIP.IPAddress, Prefix: 0, DHCP: true})
-		}
-		vminfo.IPperMac[realMAC] = entries
-	}
-}
-
-// LogMessage is an exported wrapper for logMessage that satisfies the esxissh.ProgressLogger interface.
-func (migobj *Migrate) LogMessage(message string) {
-	migobj.logMessage(message)
-}
-
-// buildProviderOptions projects vendor-specific Migrate fields into the
-// generic ProviderOptions map passed to the storage SDK. Returns nil when no
-// options apply. New vendors add their case here.
-func (migobj *Migrate) buildProviderOptions() map[string]string {
-	opts := map[string]string{}
-	if migobj.VendorType == netappsdk.VendorName {
-		if migobj.NetAppSVM != "" {
-			opts[netappsdk.OptionSVM] = migobj.NetAppSVM
-		}
-		if migobj.NetAppFlexVol != "" {
-			opts[netappsdk.OptionFlexVol] = migobj.NetAppFlexVol
-		}
-	}
-	if len(opts) == 0 {
-		return nil
-	}
-	return opts
-}
-
-// InitializeStorageProvider initializes and validates the storage provider for StorageAcceleratedCopy migration
-func (migobj *Migrate) InitializeStorageProvider(ctx context.Context) error {
-	if migobj.StorageCopyMethod != constants.StorageCopyMethod {
-		migobj.logMessage("Storage copy method is not StorageAcceleratedCopy, skipping storage provider initialization")
-		return nil
-	}
-
-	migobj.logMessage("Initializing storage provider for StorageAcceleratedCopy migration")
-
-	// Validate required credentials
-	if migobj.ArrayHost == "" {
-		return fmt.Errorf("ARRAY_HOST is required for StorageAcceleratedCopy storage migration")
-	}
-	if migobj.ArrayUser == "" {
-		return fmt.Errorf("ARRAY_USER is required for StorageAcceleratedCopy storage migration")
-	}
-	if migobj.ArrayPassword == "" {
-		return fmt.Errorf("ARRAY_PASSWORD is required for StorageAcceleratedCopy storage migration")
-	}
-
-	// Create storage access info
-	accessInfo := storage.StorageAccessInfo{
-		Hostname:            migobj.ArrayHost,
-		Username:            migobj.ArrayUser,
-		Password:            migobj.ArrayPassword,
-		SkipSSLVerification: migobj.ArrayInsecure,
-		VendorType:          migobj.VendorType,
-		ProviderOptions:     migobj.buildProviderOptions(),
-	}
-
-	// Create storage provider
-	provider, err := storage.NewStorageProvider(accessInfo.VendorType)
-	if err != nil {
-		return fmt.Errorf("failed to create storage provider: %w", err)
-	}
-
-	// Connect to storage array
-	migobj.logMessage(fmt.Sprintf("Connecting to storage array: %s", migobj.ArrayHost))
-	if err := provider.Connect(ctx, accessInfo); err != nil {
-		return fmt.Errorf("failed to connect to storage array: %w", err)
-	}
-
-	// Validate credentials
-	migobj.logMessage("Validating storage array credentials...")
-	if err := provider.ValidateCredentials(ctx); err != nil {
-		return fmt.Errorf("storage array credential validation failed: %w", err)
-	}
-
-	migobj.StorageProvider = provider
-	migobj.logMessage(fmt.Sprintf("Storage provider initialized successfully: %s", provider.WhoAmI()))
-
-	return nil
-}
-
-// LoadESXiSSHKey loads the ESXi SSH private key from the Kubernetes secret
-func (migobj *Migrate) LoadESXiSSHKey(ctx context.Context) error {
-
-	migobj.logMessage(fmt.Sprintf("Loading ESXi SSH private key from secret: %s", constants.ESXiSSHSecretName))
-
-	privateKey, err := k8sutils.GetESXiSSHPrivateKey(ctx, migobj.K8sClient, constants.ESXiSSHSecretName)
-	if err != nil {
-		return errors.Wrapf(err, "failed to load ESXi SSH private key from secret %s", constants.ESXiSSHSecretName)
-	}
-
-	migobj.ESXiSSHPrivateKey = privateKey
-	migobj.logMessage("ESXi SSH private key loaded successfully")
-
-	return nil
-}
-
-// reportStagedVolumeIDs collects the Cinder volume IDs from vminfo and patches
-// them onto Migration.Status.StagedVolumeIDs. It then sends the DataCopied
-// event message via the EventReporter channel so the controller can update the
-// migration phase. Failures are non-fatal — the caller logs a warning and
-// continues.
-func (migobj *Migrate) reportStagedVolumeIDs(ctx context.Context, vminfo vm.VMInfo) error {
-	// Collect volume IDs
-	var volumeIDs []string
-	for _, disk := range vminfo.VMDisks {
-		if disk.OpenstackVol != nil && disk.OpenstackVol.ID != "" {
-			volumeIDs = append(volumeIDs, disk.OpenstackVol.ID)
-		}
-	}
-
-	// Patch Migration.Status.StagedVolumeIDs if we have a K8s client
-	if migobj.K8sClient != nil {
-		migrationName, err := utils.GetMigrationObjectName()
-		if err != nil {
-			return errors.Wrap(err, "failed to get migration object name for staged volume IDs patch")
-		}
-		migration := &vjailbreakv1alpha1.Migration{}
-		if err := migobj.K8sClient.Get(ctx, k8stypes.NamespacedName{
-			Name:      migrationName,
-			Namespace: constants.NamespaceMigrationSystem,
-		}, migration); err != nil {
-			return errors.Wrapf(err, "failed to get migration %s to patch staged volume IDs", migrationName)
-		}
-		patch := client.MergeFrom(migration.DeepCopy())
-		migration.Status.StagedVolumeIDs = volumeIDs
-		if err := migobj.K8sClient.Status().Patch(ctx, migration, patch); err != nil {
-			return errors.Wrapf(err, "failed to patch staged volume IDs on migration %s", migrationName)
-		}
-		migobj.logMessage(fmt.Sprintf("Patched StagedVolumeIDs %v on migration %s", volumeIDs, migrationName))
-	}
-
-	// Send DataCopied event so the controller can advance the migration phase
-	migobj.logMessage(constants.EventMessageDataCopied)
 
 	return nil
 }
